% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf.R, R/enframe.R
\name{eval_cdf}
\alias{eval_cdf}
\alias{get_cdf}
\alias{enframe_cdf}
\title{(Cumulative) Distribution Function}
\usage{
eval_cdf(object, at)

get_cdf(object)

enframe_cdf(
  ...,
  at,
  arg_name = ".arg",
  fn_prefix = "cdf",
  names_repair = c("universal", "unique", "minimal")
)
}
\arguments{
\item{object}{Distribution object.}

\item{at}{Vector of values to evaluate the cdf at.}

\item{arg_name}{Name of the column containing the function arguments, in the
output data frame of \code{enframe_}.}

\item{fn_name}{Name of the column containing the function evaluations, in the
output data frame of \code{enframe_}.}
}
\value{
A vector of the evaluated cdf, in the case of \code{eval_}; a data
frame with both the argument and function evaluations, in the case of
\code{enframe_}; or a vectorized function representing the cdf, in the case
of \code{get_}.
}
\description{
Access a distribution's cumulative distribution function (cdf).
}
\examples{
d <- dst_unif(0, 4)
eval_cdf(d, at = 0:4)
enframe_cdf(d, at = 0:4)
cdf <- get_cdf(d)
cdf(0:4)
}
\seealso{
Other distributional representations: 
\code{\link{enframe_pmf}()},
\code{\link{eval_density}()},
\code{\link{eval_hazard}()},
\code{\link{eval_quantile}()},
\code{\link{eval_survival}()},
\code{\link{get_chf}()}
}
\concept{distributional representations}
