% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-stepfun.R
\name{check_left_continuous}
\alias{check_left_continuous}
\alias{check_right_continuous}
\title{Check Continuity of Step Function}
\usage{
check_left_continuous(object)

check_right_continuous(object)
}
\arguments{
\item{object}{Object of class "stepfun" to check.}
}
\value{
Single logical indicating the result.
}
\description{
Check for left and right continuity of
a step function.
}
