% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframe.R, R/eval_chf.R
\name{enframe_chf}
\alias{enframe_chf}
\alias{get_chf}
\alias{eval_chf}
\title{Cumulative Hazard Function}
\usage{
enframe_chf(..., at, arg_name = ".arg", fn_prefix = "chf", sep = "_")

get_chf(object)

eval_chf(object, at)
}
\arguments{
\item{at}{Vector of values to evaluate the cdf at.}

\item{arg_name}{Name of the column containing the function arguments, in the
output data frame of \code{enframe_}.}

\item{object}{Distribution object.}
}
\value{
A vector of the evaluated chf, in the case of \code{eval_}; a data
frame with both the argument and function evaluations, in the case of
\code{enframe_}; or a vectorized function representing the chf, in the case
of \code{get_}.
}
\description{
Access a distribution's cumulative hazard function (chf).
}
\examples{
d <- dst_unif(0, 4)
eval_chf(d, at = 0:4)
enframe_chf(d, at = 0:4)
chf <- get_chf(d)
chf(0:4)
}
\seealso{
Other distributional representations: 
\code{\link{enframe_cdf}()},
\code{\link{enframe_density}()},
\code{\link{enframe_hazard}()},
\code{\link{enframe_odds}()},
\code{\link{enframe_pmf}()},
\code{\link{enframe_quantile}()},
\code{\link{enframe_survival}()}
}
\concept{distributional representations}
