% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepfun_helpers.R
\name{consolidate_weights}
\alias{consolidate_weights}
\title{Consolidate Weights}
\usage{
consolidate_weights(y, weights)
}
\arguments{
\item{y}{Vector of outcomes.}

\item{weights}{Vector of weights, one for each of \code{y}.
These need not sum to one, but must not be negative.}
}
\value{
Data frame with the following columns:
\enumerate{
  \item \code{y}: Increasing vector of unique values of \code{y}
  having positive weight.
  \item \code{prob}: Weights corresponding to each outcome.
  \item \code{tau}: Cumulative weights; that is, \code{cumsum(prob)}.
}
}
\description{
For a vector of outcomes \code{y} with a matching vector of weights,
this function provides a single non-zero, non-NA
weight per unique value of \code{y}. Weights sum to 1.
}
