% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqdr_funs.R
\name{ddst.dst}
\alias{ddst.dst}
\title{Calculate density/mass function}
\usage{
\method{ddst}{dst}(object, x)
}
\arguments{
\item{object}{Object of class "dst"}

\item{x}{Vector of values to evaluate the density or mass function at}
}
\value{
Vector of density/mass values if \code{x} is present; otherwise,
returns a function for the pdf/pmf.
}
\description{
Calculate density/mass function
}
\seealso{
\code{\link{qdst}}, \code{\link{pdst}}, \code{\link{rdst}}
}
