% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframe.R, R/eval_density.R
\name{enframe_density}
\alias{enframe_density}
\alias{eval_density}
\alias{get_density}
\title{Probability Density Function}
\usage{
enframe_density(
  ...,
  at,
  arg_name = ".arg",
  fn_prefix = "density",
  sep = "_",
  strict = TRUE
)

eval_density(object, at, strict = TRUE)

get_density(object)
}
\arguments{
\item{at}{Vector of values to evaluate the cdf at.}

\item{arg_name}{Name of the column containing the function arguments, in the
output data frame of \code{enframe_}.}

\item{object}{Distribution object.}
}
\value{
A vector of the evaluated density, in the case of \code{eval_}; a
data frame with both the argument and function evaluations, in the case of
\code{enframe_}; or a vectorized function representing the density, in the
case of \code{get_}.
}
\description{
Access a distribution's probability density function (pdf).
}
\examples{
d <- dst_unif(0, 4)
eval_density(d, at = 0:4)
enframe_density(d, at = 0:4)
density <- get_density(d)
density(0:4)
}
\seealso{
Other distributional representations: 
\code{\link{enframe_cdf}()},
\code{\link{enframe_chf}()},
\code{\link{enframe_hazard}()},
\code{\link{enframe_odds}()},
\code{\link{enframe_pmf}()},
\code{\link{enframe_quantile}()},
\code{\link{enframe_survival}()}
}
\concept{distributional representations}
