% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discontinuities.R
\name{discontinuities.mix}
\alias{discontinuities.mix}
\title{#' @export
discontinuities.dst <- function(object) {
object[\link{"discontinuities"}]
}}
\usage{
\method{discontinuities}{mix}(object, from, to, ...)
}
\description{
#' @export
discontinuities.dst <- function(object) {
object[\link{"discontinuities"}]
}
}
