% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discontinuities.R
\name{discontinuities}
\alias{discontinuities}
\alias{aggregate_weights}
\title{Discontinuities in a Distribution}
\usage{
discontinuities(object)

aggregate_weights(y, weights)
}
\arguments{
\item{object}{A distribution object.}

\item{y}{Vector of outcomes.}

\item{weights}{Vector of weights, one for each of \code{y}.
These need not sum to one, but must not be negative.}
}
\value{
Data frame with the following columns:
\enumerate{
  \item \code{y}: Increasing vector of unique values of \code{y}
  having positive weight.
  \item \code{prob}: Weights corresponding to each outcome.
  \item \code{tau}: Cumulative weights; that is, \code{cumsum(prob)}.
}
}
\description{
Discontinuity points are the coordinates marking the
positions of step discontinuities in a
distribution. They are the upper points at
each discontinuity in the cdf, together
with the size of the discontinuity gap.
}
\details{
For a vector of outcomes \code{y} with a
matching vector of \code{weights},
\code{aggregate_weights()} provides a single non-zero, non-NA
weight per unique value of \code{y}. Weights sum to 1.
}
