% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discontinuities.R
\name{discontinuities_to_variable}
\alias{discontinuities_to_variable}
\title{Determine Variable Type from Discontinuities Data Frame}
\usage{
discontinuities_to_variable(df)
}
\arguments{
\item{df}{A data frame of discontinuities, as in
the output of \code{\link{discontinuities}}.}
}
\value{
One of \code{"continuous"},
\code{"discrete"}, or \code{"mixed"}.
}
\description{
Internal function that uses a data frame of
discontinuities to determine whether the
underlying random variable is continuous,
discrete, or mixed.
}
