% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_discrete.R
\name{next_discrete_finite}
\alias{next_discrete_finite}
\alias{prev_discrete_finite}
\alias{next_discrete_natural}
\alias{prev_discrete_natural}
\title{Helper functions for finding discrete values}
\usage{
next_discrete_finite(x, from, n = 1L)

prev_discrete_finite(x, from, n = 1L)

next_discrete_natural(from, n = 1L)

prev_discrete_natural(from, n = 1L)
}
\arguments{
\item{x}{A vector of all possible discrete values.}

\item{from}{Reference value.}

\item{n}{Number of discrete values to find.}
}
\description{
\verb{*_discrete_finite} finds discrete values from a
finite set of possibilities. \verb{*_discret_natural} finds
discrete values amongst the natural numbers (including 0).
}
