% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretes.R
\name{next_discrete}
\alias{next_discrete}
\alias{prev_discrete}
\alias{num_discretes}
\alias{has_infinite_discretes}
\title{Probing discrete values in a distribution}
\usage{
next_discrete(distribution, from, n = 1L, include_from = FALSE, ...)

prev_discrete(distribution, from, n = 1L, include_from = FALSE, ...)

num_discretes(distribution, from, to, include_from, include_to)

has_infinite_discretes(distribution, from = -Inf, to = Inf)
}
\arguments{
\item{distribution}{Distribution}

\item{from}{Reference value.}

\item{n}{Number of discrete values to find.}

\item{include_from, include_to}{Logical; should the \code{from} value be included
in the query? Should the \code{to} value?}
}
\value{
For \code{next_discrete()} and \code{prev_discrete()}, a vector of
all available discrete points satisfying the query.
If less values are available than asked
via \code{n}, only those values are returned.
If infinite values satisfy the query, an error is thrown;
\code{NaN} occurs when no one particular discrete value follows, such as
when asking for the integer that comes before infinity.

For \code{num_discretes()}, a single non-negative integer, possibly infinite.
}
\description{
\code{next_discrete()} and \code{prev_discrete()} find the \code{n} discrete values
in a distribution next to a reference point. \code{num_discretes()} finds
the number of discrete values within a range.
}
\examples{
next_discrete(dst_pois(1), from = 1.3)
prev_discrete(dst_pois(1), from = 3, n = 10)
next_discrete(dst_norm(0, 1), from = 1.3, n = 4)
}
