% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{discretise}
\alias{discretise}
\alias{discretize}
\title{Discretise a Distribution}
\usage{
discretise(
  distribution,
  breakpoints,
  midpoints = c("median", "mean", "label"),
  values,
  closed = c("right", "left")
)

discretize(
  distribution,
  breakpoints,
  midpoints = c("median", "mean", "label"),
  values,
  closed = c("right", "left")
)
}
\arguments{
\item{distribution}{Distribution to discretise.}

\item{breakpoints}{Vector of breakpoints separating each bin.}

\item{midpoints, values}{Specify the values associated with each new bin
by either specifying a computation of the \code{midpoint}, or as a vector
of \code{values} (of length one more than that of \code{breakpoints}) applied to the
bins from smallest to largest. If specified,
\code{values} will override \code{midpoint}.}

\item{returns}{A finite distribution with the specified bins.}
}
\description{
Bins a distribution by specified breakpoints. Left and right endpoints of
+/- infinity are implied.
}
\note{
\code{NA} breakpoints and \code{values} (if specified) are silently removed.
Will throw an error if the lengths of \code{values} does not end up being
one more than that of \code{breakpoints}.
}
\examples{
p1 <- dst_norm(0, 1)
p2 <- discretise(p1, breakpoints = -2:2)
p3 <- discretise(p1, breakpoints = -2:2, midpoints = "mean")
plot(p1, "cdf", from = -3, to = 3)
plot(p2, "cdf", from = -3, to = 3, add = TRUE, col = "blue", n = 1000)
plot(p3, "cdf", from = -3, to = 3, add = TRUE, col = "red", n = 1000)

discretise(dst_exp(0.1), breakpoints = numeric())

dst_norm(0, 1) \%>\%
  slice_left(-2) \%>\%
  slice_right(2) \%>\%
  discretize(-2:2)
}
