% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-discretize.R
\name{discretize}
\alias{discretize}
\title{discretize a Distribution}
\usage{
discretize(
  distribution,
  breakpoints,
  midpoints = c("median", "mean", "label"),
  values,
  closed = c("right", "left")
)
}
\arguments{
\item{distribution}{Distribution to discretize.}

\item{breakpoints}{Vector of breakpoints separating each bin.}

\item{midpoints}{Calculation for the values associated with each new bin
by either specifying a computation of the midpoint.}

\item{values}{A vector of values associated with each bin (so, length is
1 greater than the length of breakpoints). If specified,
\code{values} will override \code{midpoints}.}

\item{closed}{Should the intervals be closed on the left or on the
right? One of \code{"right"} (default) or \code{"left"}.}
}
\value{
A finite distribution with the specified bins.
}
\description{
Bins a distribution by specified breakpoints. Left and right endpoints of
+/- infinity are implied.
}
\note{
\code{NA} breakpoints and \code{values} (if specified) are silently removed.
Will throw an error if the lengths of \code{values} does not end up being
one more than that of \code{breakpoints}.
}
\examples{
p1 <- dst_norm(0, 1)
p2 <- discretize(p1, breakpoints = -2:2)
p3 <- discretize(p1, breakpoints = -2:2, midpoints = "mean")
plot(p1, "cdf", from = -3, to = 3)
plot(p2, "cdf", from = -3, to = 3, add = TRUE, col = "blue", n = 1000)
plot(p3, "cdf", from = -3, to = 3, add = TRUE, col = "red", n = 1000)

discretize(dst_exp(0.1), breakpoints = numeric())

dst_norm(0, 1) |>
  slice_left(-2) |>
  slice_right(2) |>
  discretize(-2:2)
}
