% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots_to_dsts.R
\name{dots_to_dsts}
\alias{dots_to_dsts}
\title{Convert Distributions in Ellipsis to List}
\usage{
dots_to_dsts(..., na.rm = FALSE)
}
\arguments{
\item{...}{Distribution objects, possibly also lists of distributions.}

\item{na.rm}{Logical; remove NA entries? Note that NULL entries are
always removed.}
}
\value{
A list of distributions contained in the \code{...}, with NULL
entries discarded. If no distributions are present, returns \code{list()}.
}
\description{
Flattens distributions placed in an ellipsis argument into a list,
so that the ellipsis can include distributions themselves and lists
of distributions.
}
\details{
An error is thrown if, after discarding NULL entries,
\code{...} contains non-distributions. This function is essentially a
wrapper around \code{rlang::flatten()}.
}
\examples{
d <- dst_norm(0, 1)
distplyr:::dots_to_dsts(d, list(d, d), NULL)
}
