% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst.R
\name{dst}
\alias{dst}
\title{Make a distribution}
\usage{
dst(fun_cumu, fun_quant, fun_prob, fun_rand, name = NULL, param = NULL,
  prop = NULL)
}
\arguments{
\item{fun_cumu, fun_quant, fun_prob, fun_rand}{cdf, quantile function, density/mass
function, and random number generator for a distribution.}

\item{name}{A name for the distribution (such as a parametric family name)}

\item{param}{Parameters for the distribution, if parameteric.}

\item{prop}{Properties of the distribution, such as mean, variance,
EVI, etc. (of your choosing).}
}
\value{
An object of class "dst", which (for now) is a list holding
the above arguments, including survival and hazard functions.
}
\description{
Make a distribution object.
}
