% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_beta.R
\name{dst_beta}
\alias{dst_beta}
\title{Beta Distribution}
\usage{
dst_beta(shape1, shape2)
}
\arguments{
\item{shape1, shape2}{Parameters, greater than 0. Also known as alpha
and beta.}
}
\description{
Makes a distribution belonging to the family of
beta distributions.
}
\examples{
dst_beta(2, 3)
}
