% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_degen.R
\name{dst_degen}
\alias{dst_degen}
\title{Make a Degenerate Distribution}
\usage{
dst_degen(loc)
}
\arguments{
\item{loc}{Parameter of the distribution family.}
}
\value{
Object of class "dst".
}
\description{
Makes a distribution belonging to the degenerate family of
distributions. That is, distributions of fixed values.
}
