% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_emp.R
\name{dst_emp}
\alias{dst_emp}
\title{Make an Empirical distribution}
\usage{
dst_emp(x)
}
\arguments{
\item{x}{Vector of values to construct empirical distribution from.}
}
\value{
Object of class "dst".
}
\description{
Makes an empirical distribution from a sample of univariate
data.
}
\seealso{
\code{\link{eqf}}, \code{\link{epmf}}
}
