% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_empirical.R
\name{dst_empirical}
\alias{dst_empirical}
\title{Create an Empirical Distribution}
\usage{
dst_empirical(y, data, weights = 1, ...)
}
\arguments{
\item{y}{<\code{data-masking}>
Outcomes to comprise the distribution. Should either
evaluate to an (atomic) vector, or be a name in the specified data.}

\item{data}{Data frame containing the outcomes \code{y} and/or
\code{weights}. Optional.}

\item{weights}{<\code{data-masking}>
Weights to assign each outcome in \code{y}. Will be
normalized so that the weights add up to 1
(unlike \code{\link[=dst_finite]{dst_finite()}}),
representing probabilities.}

\item{...}{Additional arguments, currently not used.}
}
\value{
An object of class \code{c("finite", "dst")}.
}
\description{
An empirical distribution is a non-parametric way to
estimate a distribution using data. By default,
it assigns equal probability to all observations
(this can be overridden with the \code{weights} argument).
Identical to \code{\link[=dst_finite]{dst_finite()}} with weights as probabilities,
except weights need not add to 1.
}
\examples{
require(graphics)
require(datasets)
marg <- dst_empirical(hp, data = mtcars)
plot(marg, "cdf", n = 1001)

K <- function(x) dnorm(x, sd = 25)
cond <- dst_empirical(hp, data = mtcars, weights = K(disp - 150))
plot(cond, "cdf", n = 1001, lty = 2, add = TRUE)
}
\seealso{
\code{\link[=dst_finite]{dst_finite()}}
}
