% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_finite.R
\name{dst_finite}
\alias{dst_finite}
\title{Create a Finite Distribution}
\usage{
dst_finite(y, probs, data, ...)
}
\arguments{
\item{y}{<\code{data-masking}>
Outcomes to comprise the distribution. Should either
evaluate to an (atomic) vector, or be a name in the specified data.}

\item{probs}{<\code{data-masking}>
Probabilities corresponding to the outcomes in \code{y}.
Must not be negative, but \strong{must sum to 1}
(unlike \code{\link[=dst_empirical]{dst_empirical()}}). Should either
evaluate to a vector, or be a name in the specified data.}

\item{data}{Data frame containing the outcomes \code{y} and/or
probabilities \code{probs}. Optional.}

\item{...}{Additional arguments, currently not used.}
}
\value{
An object of class \code{c("finite", "dst")}.
}
\description{
A finite distribution assigns probabilities to a finite
collection of values. This includes categorical distributions.
}
\note{
This distribution is called "finite" and not
"discrete", because a discrete distribution could have
an infinite amount of possible outcomes, as in the
Poisson distribution.
}
\examples{
dst_finite(1:5, probs = rep(0.2, 5))
}
\seealso{
\code{\link[=dst_empirical]{dst_empirical()}}
}
