% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_gpd.R
\name{dst_gpd}
\alias{dst_gpd}
\title{Generalized Pareto Distribution}
\usage{
dst_gpd(loc, scale, shape)
}
\arguments{
\item{loc, scale, shape}{Parameters of the GPD.}
}
\value{
Object of class "dst" of a GPD.
}
\description{
Makes a distribution belonging to the family of
generalized Pareto distributions (GPD).
}
