% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_lnorm.R
\name{dst_lnorm}
\alias{dst_lnorm}
\title{Log Normal Distribution}
\usage{
dst_lnorm(meanlog, variancelog)
}
\arguments{
\item{meanlog, variancelog}{Mean and variance of the distribution
on a log scale.}
}
\value{
Object of class \code{c("lnorm", "dst")}.
dst_lnorm(0, 1)
}
\description{
Makes a distribution belonging to the family of
Log Normal distributions.
}
