% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_norm.R
\name{dst_norm}
\alias{dst_norm}
\title{Normal (Gaussian) Distribution}
\usage{
dst_norm(mean, variance)
}
\arguments{
\item{mean, variance}{Mean and variance of the distribution.}
}
\value{
Object of class "dst".
}
\description{
Makes a distribution belonging to the family of
Normal (Gaussian) distributions.
}
