% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_parametric.R
\name{dst_parametric}
\alias{dst_parametric}
\title{Convert base distribution to distplyr}
\usage{
dst_parametric(
  .name,
  ...,
  .variable = c("unknown", "continuous", "discrete", "mixed")
)
}
\arguments{
\item{.name}{Name of the distribution. Specifically, a string with
the suffix of the \code{p}, \code{d}, \code{q}, \code{r} functions (such as \code{"norm"}).}

\item{...}{Name-value pairs of parameters. Names must be found in the
parameter names of the \code{p}, \code{d}, \code{q}, \code{r} functions.}

\item{.variable}{Type of random variable represented by the distribution.
Warning: defaults to "unknown", where density, pmf, and hazard cannot be
evaluated.}
}
\description{
When a distribution has \code{p}, \code{d}, \code{q}, \code{r} functions available (such as
\code{pnorm()}, \code{dnorm()}, etc.), \code{dst_parametric()} creates a distribution
that draws on these four functions.
}
\examples{
d <- dst_parametric("norm", mean = 0, sd = 1, .variable = "continuous")
eval_density(d, at = -3:3)
eval_pmf(d, at = -3:3, strict = FALSE)
eval_hazard(d, at = -3:3)
}
