% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_pois.R
\name{dst_pois}
\alias{dst_pois}
\title{Poisson Distribution}
\usage{
dst_pois(lambda)
}
\arguments{
\item{lambda}{Mean of the Poisson distribution.}
}
\value{
Object of class "dst".
dst_pois(1)
}
\description{
Makes a distribution belonging to the family of
Poisson distributions.
}
