% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_unif.R
\name{dst_unif}
\alias{dst_unif}
\title{Make a Uniform distribution}
\usage{
dst_unif(min, max)
}
\arguments{
\item{min, max}{Parameters of the distribution family. Need not specify
anything (uses tidy evaluation).}
}
\value{
Object of class "dst".
dst_unif(0, 1)
}
\description{
Makes a distribution belonging to the continuous uniform family of
distributions. Note: this distribution has been adapted to trial
a tidy evaluation framework.
}
