% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-stepfun.R
\name{eqf}
\alias{eqf}
\alias{epmf}
\title{Empirical Functions for Distributions}
\usage{
eqf(x)

epmf(x)
}
\arguments{
\item{x}{Numeric vector of the observations for which to make the quantile
function}
}
\value{
Vectorized functions; quantile function for \code{eqf()}, and
probability mass function for \code{epmf()}.
}
\description{
Makes empirical mass and quantile functions from a univariate
sample, just as \code{stats::ecdf()} makes empirical cdf's.
}
