% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqf.R
\name{eqf}
\alias{eqf}
\title{Empirical quantile function}
\usage{
eqf(x)
}
\arguments{
\item{x}{Numeric vector of the observations for which to make the quantile
function}
}
\description{
A friend of \code{ecdf}, this function returns a (vectorized) empirical quantile
function that's the inverse of ecdf.
}
