% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_fn.R
\name{eval_cdf}
\alias{eval_cdf}
\title{Evaluate cdf}
\usage{
eval_cdf(object, at)
}
\arguments{
\item{object}{Object of class "dst"}

\item{at}{Vector of values to evaluate the cdf at}
}
\value{
Vector of cdf values
}
\description{
Evaluate cdf
}
\seealso{
\code{\link{eval_quantfn}},
\code{\link{eval_probfn}},
\code{\link{eval_randfn}},
\code{\link{eval_hazfn}},
\code{\link{eval_surv}}
}
