% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_fn-generics.R
\name{eval_hazard}
\alias{eval_hazard}
\title{Evaluate Hazard Function}
\usage{
eval_hazard(object, at)
}
\arguments{
\item{object}{Object of class "dst"}

\item{at}{Vector of values to evaluate the hazard function at.}
}
\value{
Vector of the evaluated hazard function
}
\description{
Evaluate Hazard Function
}
\seealso{
\code{\link{eval_quantile}},
\code{\link{eval_probfn}},
\code{\link{eval_cdf}},
\code{\link{eval_randfn}},
\code{\link{eval_survival}}
}
