% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_fn.R
\name{eval_hazfn}
\alias{eval_hazfn}
\title{Evaluate Hazard Function}
\usage{
eval_hazfn(object, at)
}
\arguments{
\item{object}{Object of class "dst"}

\item{at}{Vector of values to evaluate the hazard function at.}
}
\value{
Vector of the evaluated hazard function
}
\description{
Evaluate Hazard Function
}
\seealso{
\code{\link{eval_quantfn}},
\code{\link{eval_probfn}},
\code{\link{eval_cdf}},
\code{\link{eval_randfn}},
\code{\link{eval_surv}}
}
