% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_fn.R
\name{eval_probfn}
\alias{eval_probfn}
\title{Evaluate probability density/mass function}
\usage{
eval_probfn(object, at)
}
\arguments{
\item{object}{Object of class "dst"}

\item{at}{Vector of values to evaluate the density or mass function at}
}
\value{
Vector of density/mass values
}
\description{
Evaluate probability density/mass function
}
\seealso{
\code{\link{eval_quantfn}},
\code{\link{eval_cdf}},
\code{\link{eval_randfn}},
\code{\link{eval_hazfn}},
\code{\link{eval_surv}}
}
