% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_fn.R
\name{eval_quantfn}
\alias{eval_quantfn}
\title{Evaluate quantile function}
\usage{
eval_quantfn(object, at)
}
\arguments{
\item{object}{Object of class "dst"}

\item{at}{Vector of values to evaluate the quantile function at}
}
\value{
Vector of quantiles
}
\description{
Evaluate quantile function
}
\seealso{
\code{\link{eval_probfn}},
\code{\link{eval_cdf}},
\code{\link{eval_randfn}},
\code{\link{eval_hazfn}},
\code{\link{eval_surv}}
}
