% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graft-old.R
\name{eval_quantile_at_left}
\alias{eval_quantile_at_left}
\title{#' Graft Distributions
#' Replace the tail of a distribution
#' @export
graft_right <- function(object, dst_right, at, include_at = TRUE, ...) {
if (!is_graft(object)) {
graft(
object,
NULL,
dst_right,
NULL,
at,
NULL,
include_at,
dst_right,
NULL
)
} else if (is.null(object$components$dst_left) != TRUE) {
graft(
object$component$base,
object$component$dst_left,
dst_right,
object$component$left_at,
at,
object$component$include_at_left_in_base,
include_at,
dst_right,
dst_left
)
} else if (is.null(object$components$dst_left) == TRUE) {
graft(
object$component$base,
NULL,
dst_right,
NULL,
at,
NULL,
include_at,
dst_right,
NULL
)
}
}
#' @export
graft_left <- function(object, dst_left, at, include_at = TRUE, ...) {
if (!is_graft(object)) {
graft(
object,
dst_left,
NULL,
at,
NULL,
include_at,
NULL,
dst_left,
NULL
)
} else if (is.null(object$components$dst_right) != TRUE) {
graft(
object$component$base,
dst_left,
object$component$dst_right,
at,
object$component$right_at,
include_at,
dst_left,
dst_right
)
} else if (is.null(object$components$dst_right) == TRUE) {
graft(
object$component$base,
dst_left,
NULL,
at,
NULL,
include_at,
NULL,
dst_left,
NULL
)
}
}}
\usage{
eval_quantile_at_left(object, at)
}
\description{
#' Graft Distributions
#' Replace the tail of a distribution
#' @export
graft_right <- function(object, dst_right, at, include_at = TRUE, ...) {
if (!is_graft(object)) {
graft(
object,
NULL,
dst_right,
NULL,
at,
NULL,
include_at,
dst_right,
NULL
)
} else if (is.null(object$components$dst_left) != TRUE) {
graft(
object$component$base,
object$component$dst_left,
dst_right,
object$component$left_at,
at,
object$component$include_at_left_in_base,
include_at,
dst_right,
dst_left
)
} else if (is.null(object$components$dst_left) == TRUE) {
graft(
object$component$base,
NULL,
dst_right,
NULL,
at,
NULL,
include_at,
dst_right,
NULL
)
}
}
#' @export
graft_left <- function(object, dst_left, at, include_at = TRUE, ...) {
if (!is_graft(object)) {
graft(
object,
dst_left,
NULL,
at,
NULL,
include_at,
NULL,
dst_left,
NULL
)
} else if (is.null(object$components$dst_right) != TRUE) {
graft(
object$component$base,
dst_left,
object$component$dst_right,
at,
object$component$right_at,
include_at,
dst_left,
dst_right
)
} else if (is.null(object$components$dst_right) == TRUE) {
graft(
object$component$base,
dst_left,
NULL,
at,
NULL,
include_at,
NULL,
dst_left,
NULL
)
}
}
}
