% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-quantile.R
\name{eval_quantile_from_cdf}
\alias{eval_quantile_from_cdf}
\title{Evaluate Quantiles using a CDF}
\usage{
eval_quantile_from_cdf(cdf, discon, at, tol, maxiter)
}
\arguments{
\item{cdf}{Function representing the cdf.}

\item{discon}{A data frame of discontinuities, as in the
output of \code{discontinuities()}.}

\item{at}{A vector of values for which to evaluate the
quantile function.}

\item{tol, maxiter}{Tolerance and maximum number of iterations}
}
\description{
Intended for internal use only.
}
