% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_fn.R
\name{eval_surv}
\alias{eval_surv}
\title{Evaluate Survival Function}
\usage{
eval_surv(object, at)
}
\arguments{
\item{object}{Object of class "dst"}

\item{at}{Vector of values to evaluate the survival function at.}
}
\value{
Vector of the evaluated survival function
}
\description{
Evaluate Survival Function
}
\seealso{
\code{\link{eval_quantfn}},
\code{\link{eval_probfn}},
\code{\link{eval_cdf}},
\code{\link{eval_hazfn}},
\code{\link{eval_randfn}}
}
