% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation-survival.R
\name{eval_survival}
\alias{eval_survival}
\title{Evaluate Survival Function}
\usage{
eval_survival(object, at)
}
\arguments{
\item{object}{Object of class "dst"}

\item{at}{Vector of values to evaluate the survival function at.}
}
\value{
Vector of the evaluated survival function
}
\description{
Evaluate Survival Function
}
\seealso{
\code{\link{eval_quantile}},
\code{\link{eval_probfn}},
\code{\link{eval_cdf}},
\code{\link{eval_hazard}},
\code{\link{eval_randfn}}
}
