% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evi.R
\name{evi.dst}
\alias{evi.dst}
\title{Get the EVI of a distribution}
\usage{
\method{evi}{dst}(object)
}
\arguments{
\item{object}{Object of class "dst" to obtain mean from.}
}
\value{
A single numeric
}
\description{
Extreme value index
}
\details{
Doesn't calculate if not present.
}
