% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evi.R
\name{evi}
\alias{evi}
\title{Get the EVI of a distribution}
\usage{
evi(object)
}
\arguments{
\item{object}{Object of class "dst" to obtain EVI from.}
}
\value{
A single numeric
}
\description{
EVI = Extreme value index
}
\details{
Doesn't calculate EVI if not present in the object.
}
