% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-exp.R
\name{exp_distribution}
\alias{exp_distribution}
\title{Exponential Transformation of a Distribution}
\usage{
exp_distribution(distribution)
}
\arguments{
\item{distribution}{A probability distribution.}
}
\value{
A distribution transformed by the exponential function.
Specifically, a distribution with subclass "exponential".
}
\description{
Apply the exponential function to a distribution. Specifically, if \code{X}
is a random variable coming from a distribution, \code{exp_distribution()}
returns the distribution of \code{exp(X)}.
}
\examples{
exp_distribution(distionary::dst_norm(0, 1))
}
