% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexible_cdf.R
\name{prob_left}
\alias{prob_left}
\title{Find the probability left or right of a number}
\usage{
prob_left(distribution, of, inclusive)
}
\arguments{
\item{distribution}{Distribution to find probabilities of.}

\item{of}{Find the probability to the left or right \emph{of} this number.
Could be a vector.}

\item{inclusive}{Should \code{of} be included in the probability calculation?
Logical.}
}
\description{
Almost a direct application of cdf and survival function,
but allows for the cdf to be calculated using \code{<} and the survival
function to be calculated using \code{>=}.
}
