% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqdr_funs.R
\name{fun_cumu}
\alias{fun_cumu}
\title{Calculate cdf}
\usage{
fun_cumu(object, at)
}
\arguments{
\item{object}{Object of class "dst"}

\item{at}{Vector of values to evaluate the cdf at}
}
\value{
Vector of cdf values if \code{at} is present; otherwise,
returns a function for the cdf.
}
\description{
Calculate cdf
}
\seealso{
\code{\link{fun_quant}},
\code{\link{fun_prob}},
\code{\link{fun_rand}},
\code{\link{fun_haz}},
\code{\link{fun_surv}}
}
