% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqdr_funs.R
\name{fun_haz}
\alias{fun_haz}
\title{Calculate Hazard Function}
\usage{
fun_haz(object, at)
}
\arguments{
\item{object}{Object of class "dst"}

\item{at}{Vector of values to evaluate the hazard function at (optional).}
}
\value{
Vector of the evaluated hazard function
if \code{at} is present; otherwise,
returns a function for the hazard.
}
\description{
Calculate Hazard Function
}
\seealso{
\code{\link{fun_quant}},
\code{\link{fun_prob}},
\code{\link{fun_cumu}},
\code{\link{fun_rand}},
\code{\link{fun_surv}}
}
