% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqdr_funs.R
\name{fun_prob}
\alias{fun_prob}
\title{Calculate density/mass function}
\usage{
fun_prob(object, at)
}
\arguments{
\item{object}{Object of class "dst"}

\item{at}{Vector of values to evaluate the density or mass function at}
}
\value{
Vector of density/mass values if \code{at} is present; otherwise,
returns a function for the pdf/pmf.
}
\description{
Calculate density/mass function
}
\seealso{
\code{\link{fun_quant}},
\code{\link{fun_cumu}},
\code{\link{fun_rand}},
\code{\link{fun_haz}},
\code{\link{fun_surv}}
}
