% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqdr_funs.R
\name{fun_quant}
\alias{fun_quant}
\title{Calculate quantile function}
\usage{
fun_quant(object, at)
}
\arguments{
\item{object}{Object of class "dst"}

\item{at}{Vector of values to evaluate the quantile function at}
}
\value{
Vector of quantiles if \code{at} is present; otherwise,
returns a function for the quantile function.
}
\description{
Calculate quantile function
}
\seealso{
\code{\link{fun_prob}},
\code{\link{fun_cumu}},
\code{\link{fun_rand}},
\code{\link{fun_haz}},
\code{\link{fun_surv}}
}
