% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqdr_funs.R
\name{fun_rand}
\alias{fun_rand}
\title{Generate random sample from a distribution}
\usage{
fun_rand(object, at)
}
\arguments{
\item{object}{Object of class "dst"}

\item{at}{Number of observations to generate}
}
\value{
Vector of independent values drawn from the inputted distribution
if \code{at} is present; otherwise,
returns a function for a random number generator.
}
\description{
Generate random sample from a distribution
}
\seealso{
\code{\link{fun_quant}},
\code{\link{fun_prob}},
\code{\link{fun_cumu}},
\code{\link{fun_haz}},
\code{\link{fun_surv}}
}
