% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqdr_funs.R
\name{fun_surv}
\alias{fun_surv}
\title{Calculate Survival Function}
\usage{
fun_surv(object, at)
}
\arguments{
\item{object}{Object of class "dst"}

\item{at}{Vector of values to evaluate the survival function at (optional).}
}
\value{
Vector of the evaluated survival function
if \code{at} is present; otherwise,
returns a function for the survival.
}
\description{
Calculate Survival Function
}
\seealso{
\code{\link{fun_quant}},
\code{\link{fun_prob}},
\code{\link{fun_cumu}},
\code{\link{fun_haz}},
\code{\link{fun_rand}}
}
