% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn.R
\name{get_cdf}
\alias{get_cdf}
\title{Get cdf}
\usage{
get_cdf(object)
}
\arguments{
\item{object}{Object of class "dst"}
}
\value{
a vectorized function of the cdf
}
\description{
Get cdf
}
\seealso{
\code{\link{get_quantfn}},
\code{\link{get_probfn}},
\code{\link{get_randfn}},
\code{\link{get_hazfn}},
\code{\link{get_surv}}
}
