% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_evi.R
\name{get_evi}
\alias{get_evi}
\title{Get the EVI of a distribution}
\usage{
get_evi(object)
}
\arguments{
\item{object}{Distribution object to obtain EVI from.}
}
\value{
A single numeric
}
\description{
EVI = Extreme value index
}
\details{
Doesn't calculate EVI if not present in the object.
}
\examples{
get_evi(dst_gpd(0, 1, 2))
get_evi(dst_gpd(0, 1, -1))
get_evi(dst_unif(0, 10))
get_evi(dst_norm(5, 5))
get_evi(stepdst(1:10))
}
