% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn.R
\name{get_hazfn}
\alias{get_hazfn}
\title{Get Hazard Function}
\usage{
get_hazfn(object)
}
\arguments{
\item{object}{Object of class "dst"}
}
\value{
a vectorized function of the hazard.
}
\description{
Get Hazard Function
}
\seealso{
\code{\link{get_quantfn}},
\code{\link{get_probfn}},
\code{\link{get_cdf}},
\code{\link{get_randfn}},
\code{\link{get_surv}}
}
