% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean.R
\name{get_mean}
\alias{get_mean}
\title{Mean of a Distribution}
\usage{
get_mean(object, ...)
}
\arguments{
\item{object}{A distribution object from which to obtain the mean.}

\item{...}{Arguments to pass to the \code{integrate()} function (if needed).}
}
\value{
A single numeric.
}
\description{
Get the mean of a distribution.
}
\details{
If the mean is not already available in the distribution
object, and if the distribution is not a step distribution,
the mean is calculated using the \code{stats::integrate()}
function.
}
