% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation-probfn.R
\name{get_probfn}
\alias{get_probfn}
\title{Get probability density/mass function}
\usage{
get_probfn(object)
}
\arguments{
\item{object}{Object of class "dst"}
}
\value{
a vectorized function of the pdf/pmf.
}
\description{
Get probability density/mass function
}
\seealso{
\code{\link{get_quantile}},
\code{\link{get_cdf}},
\code{\link{get_randfn}},
\code{\link{get_hazard}},
\code{\link{get_survival}}
}
