% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn.R
\name{get_quantfn}
\alias{get_quantfn}
\title{Get quantile function}
\usage{
get_quantfn(object)
}
\arguments{
\item{object}{Object of class "dst"}
}
\value{
a vectorized function of the quantile function.
}
\description{
Get quantile function
}
\seealso{
\code{\link{get_probfn}},
\code{\link{get_cdf}},
\code{\link{get_randfn}},
\code{\link{get_hazfn}},
\code{\link{get_surv}}
}
