% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation-randfn.R
\name{get_randfn}
\alias{get_randfn}
\title{Random sampler function from a distribution}
\usage{
get_randfn(object)
}
\arguments{
\item{object}{Object of class "dst"}
}
\value{
a function accepting an integer number of values to generate.
}
\description{
Random sampler function from a distribution
}
\seealso{
\code{\link{get_quantile}},
\code{\link{get_probfn}},
\code{\link{get_cdf}},
\code{\link{get_hazard}},
\code{\link{get_survival}}
}
