% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn.R
\name{get_surv}
\alias{get_surv}
\title{Calculate Survival Function}
\usage{
get_surv(object)
}
\arguments{
\item{object}{Object of class "dst"}
}
\value{
a vectorized function of the survival.
}
\description{
Calculate Survival Function
}
\seealso{
\code{\link{get_quantfn}},
\code{\link{get_probfn}},
\code{\link{get_cdf}},
\code{\link{get_hazfn}},
\code{\link{get_randfn}}
}
