% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn-generics.R
\name{get_survival}
\alias{get_survival}
\title{Survival Function}
\usage{
get_survival(object)
}
\arguments{
\item{object}{Object of class "dst"}
}
\value{
a vectorized function of the survival.
}
\description{
Survival Function
}
\seealso{
\code{\link{get_quantile}},
\code{\link{get_probfn}},
\code{\link{get_cdf}},
\code{\link{get_hazard}},
\code{\link{get_randfn}}
}
