% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variance.R
\name{get_variance}
\alias{get_variance}
\alias{get_sd}
\title{Variance/Standard Deviation of a Distribution}
\usage{
get_variance(object, ...)

get_sd(object, ...)
}
\arguments{
\item{object}{A distribution object.}

\item{...}{Arguments to pass to the \code{integrate()} function (if needed).}
}
\value{
A single numeric
}
\description{
Get the mean and standard deviation of a distribution.
}
\details{
If the statistic is not already available in the distribution
object, and if the distribution is not a step distribution,
the variance/standard deviation is calculated using the
\code{stats::integrate()} function.
}
