% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graft.R
\name{graft_right}
\alias{graft_right}
\title{Graft Distributions
Replace the tail of a distribution}
\usage{
graft_right(object, dst_right, at, include_at = TRUE, ...)
}
\description{
Graft Distributions
Replace the tail of a distribution
}
