% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_pdf_pmf.R
\name{has_pmf}
\alias{has_pmf}
\alias{has_pdf}
\title{Check if a distribution has a pmf/pdf}
\usage{
has_pmf(...)

has_pdf(...)
}
\arguments{
\item{object}{Object of class "dst"}
}
\value{
Logical value
}
\description{
Check if a distribution has a pmf/pdf
}
