% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_fn-generics.R, R/get_fn-generics.R
\name{eval_hazard}
\alias{eval_hazard}
\alias{eval_chf}
\alias{get_chf}
\alias{get_hazard}
\title{Evaluate Hazard Function}
\usage{
eval_hazard(object, at)

eval_chf(object, at)

get_chf(object)

get_hazard(object)
}
\arguments{
\item{object}{Distribution object}

\item{at}{Vector of values to evaluate the hazard function at.}
}
\value{
Vector of the evaluated hazard function
}
\description{
Evaluate hazard function or cumulative
hazard function (chf).

Hazard and cumulative hazard functions.
}
\seealso{
\code{\link{eval_quantile}},
\code{\link{eval_probfn}},
\code{\link{eval_cdf}},
\code{\link{eval_randfn}},
\code{\link{eval_survival}}
}
