% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifications-finite.R
\name{invert.finite}
\alias{invert.finite}
\title{Inverse Distribution}
\usage{
invert.finite(distribution)
}
\arguments{
\item{distribution}{A probability distribution.}
}
\value{
An inverse distribution.
Specifically, a distribution with subclass "inverse".
}
\description{
Access the inverse of a distribution. Specifically, if \code{X} is a random
variable coming from a distribution, \code{invert()} returns the
distribution of \code{1 / X}.
}
\examples{
1 / (distionary::dst_pois(3.4) + 1)
invert(distionary::dst_norm(0, 1))
}
\seealso{
\code{flip()}, \code{scale()}
}
