% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_continuity.R
\name{is_left_continuous}
\alias{is_left_continuous}
\alias{is_right_continuous}
\title{Check Continuity of Step Function}
\usage{
is_left_continuous(object)

is_right_continuous(object)
}
\arguments{
\item{object}{Object of class "stepfun" to check.}
}
\value{
Single logical indicating the result.
}
\description{
Check for left and right continuity of
a step function.
}
