% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discontinuities.R
\name{is_discontinuities_df}
\alias{is_discontinuities_df}
\title{Check if a Data Frame is a Discontinuity Data Frame}
\usage{
is_discontinuities_df(df)
}
\arguments{
\item{df}{Data frame to check}
}
\value{
Logical.
}
\description{
The official names in a discontinuity data frame are
determined (and checked) here.
}
