% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kurtosis.R
\name{get_kurtosis_raw}
\alias{get_kurtosis_raw}
\alias{get_kurtosis_exc}
\title{Kurtosis}
\usage{
get_kurtosis_raw(object)

get_kurtosis_exc(object)
}
\arguments{
\item{object}{Distribution object}
}
\description{
Get the kurtosis (raw) or excess kurtosis (exc) of a distribution.
}
