% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-flip.R, R/verb-multiply.R, R/verb-shift.R
\name{flip}
\alias{flip}
\alias{multiply}
\alias{shift}
\title{Linear Transformations}
\usage{
flip(distribution)

multiply(distribution, constant)

shift(distribution, constant)
}
\arguments{
\item{distribution}{A probability distribution of class \code{"dst"}.}

\item{constant}{A single numeric by which to shift or multiply the
distribution by.}
}
\value{
A distribution, shifted or multiplied by the constant.
Specifically, a distribution with subclass "shift", "scale", or "flip".
}
\description{
Change the location (\code{shift()}) and scale (\code{multiply()}) of a distribution.
The \code{flip()} verb is also available for negating a distribution, and is
called internally when \code{multiply()} receives a negative constant. These
transformations can also be accessed by the binary operators \code{+}, \code{-},
and \code{*}.
}
\details{
If \code{X} is a random variable,
then the transformations correspond to the following:
\itemize{
\item \code{shift()} provides the distribution of \code{X + constant}.
\item \code{multiply()} provides the distribution of \code{X * constant}.
\item \code{flip()} provides the distribution of \code{-X}.
}

Simplifications are made in some situations, and will improve with future
versions of distplyr.
}
\examples{
d_pois <- distionary::dst_pois(1.1)
d_norm <- distionary::dst_norm(4, 1)
d_unif <- distionary::dst_unif(0, 1)

# Shift a Poisson distribution by 1.
shift(d_pois, 1)
d_pois + 1

# Multiply a Uniform distribution by 2: simplifies to a Uniform distribution.
multiply(d_unif, 2)
d_unif * 2

# Flip a Normal distribution: simplifies to another Normal distribution.
flip(d_norm)
-d_norm

# Combine multiple operations:
4 - 2 * d_pois
}
\seealso{
\code{\link[=invert]{invert()}} for division by a distribution.
}
