% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-log.R
\name{log_distribution}
\alias{log_distribution}
\title{Logarithmic Transformation of a Distribution}
\usage{
log_distribution(distribution, base = exp(1))
}
\arguments{
\item{distribution}{A probability distribution.}

\item{base}{A positive numeric value specifying the base of the logarithm.
Defaults to Euler's constant (natural logarithm).}
}
\value{
A distribution transformed by the logarithm function.
Specifically, a distribution with subclass "logarithmic".
}
\description{
Apply the logarithm function to a distribution. Specifically, if \code{X}
is a random variable coming from a distribution, \code{log_distribution()}
returns the distribution of \code{log(X)} with a specified base.
}
\note{
An error is returned if the original distribution
has non-positive values as possible outcomes.
}
\examples{
log_distribution(distionary::dst_unif(1, 10)) # Natural log
log_distribution(distionary::dst_unif(1, 10), base = 10) # Log base 10
}
