% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discontinuities.R
\name{make_discontinuities_df}
\alias{make_discontinuities_df}
\title{Make a Data Frame of Discontinuities}
\usage{
make_discontinuities_df(location, size)
}
\arguments{
\item{location, size}{Vectors of equal length from which
to construct a data frame.}
}
\description{
Places the components of step discontinuities
into a data frame, and checks that the result is named
appropriately.
}
