% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discontinuities.R
\name{make_empty_discontinuities_df}
\alias{make_empty_discontinuities_df}
\title{Rowless Step Discontinuity Data Frame}
\usage{
make_empty_discontinuities_df()
}
\description{
Internal function for making a step
discontinuity data frame having no
rows, useful for distributions that
do not have any step discontinuities
(i.e., continuous distributions)
}
