% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-max.R
\name{maximise}
\alias{maximise}
\title{Max Value of Several Distributions}
\usage{
maximise(...)
}
\arguments{
\item{...}{Distribution objects}
}
\value{
A distribution of class \code{"max"}.
}
\description{
For a collection of distributions, this function provides the
distribution of the maximum value from independent draws of
each component distribution.
}
\details{
To use precise language, if \code{X1}, ..., \code{Xp} are
\code{p} independent random variables corresponding to the distributions
in \code{...}, then the distribution returned is of \code{max(X1, ..., Xp)}.
}
