% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-maximize.R
\name{maximize}
\alias{maximize}
\alias{maximise}
\title{Max Value of Several Distributions}
\usage{
maximize(
  ...,
  draws = 1,
  na_action_dst = c("null", "drop", "fail"),
  na_action_draws = c("null", "drop", "fail")
)

maximise(..., draws = 1)
}
\arguments{
\item{...}{Distribution objects, or lists of distributions.}

\item{draws}{Number of draws from each distribution considered in the
maximum. Either a single numeric applying to all distributions in \code{...},
or a vector matching the number of distributions in \code{...}.}
}
\value{
A distribution of class \code{"max"}.
}
\description{
For a collection of distributions, this function provides the
distribution of the maximum value from independent draws of
each component distribution.
}
\details{
To use precise language, if \code{X1}, ..., \code{Xp} are
\code{p} independent random variables corresponding to the distributions
in \code{...}, then the distribution returned is of \code{max(X1, ..., Xp)}.
Distributions and \code{draws} are recycled to the same length using
\code{vctrs::vec_recycle_common()}.
}
