% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean.dst}
\alias{mean.dst}
\title{Get the mean of a distribution}
\usage{
\method{mean}{dst}(object, ..., verbose = FALSE)
}
\arguments{
\item{object}{Object of class "dst" to obtain mean from.}

\item{...}{Arguments to pass to the \code{integrate()} function.}

\item{verbose}{Print output of \code{integrate()} function?}
}
\value{
A single numeric
}
\description{
Get the mean of a distribution
}
\details{
If the mean is not already available in the distribution
object, this is calculated using the \code{stats::integrate()}
function.
}
