% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median.R
\name{median.dst}
\alias{median.dst}
\title{Median of a Distribution}
\usage{
\method{median}{dst}(object, ...)
}
\arguments{
\item{object}{Distribution object}

\item{...}{Other arguments to pass to specific methods}
}
\description{
Median of a Distribution
}
