% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosis.R, R/mean.R, R/median.R, R/skewness.R,
%   R/stdev.R, R/variance.R
\name{kurtosis_raw}
\alias{kurtosis_raw}
\alias{kurtosis_exc}
\alias{mean.dst}
\alias{median.dst}
\alias{skewness}
\alias{stdev}
\alias{stdev.dst}
\alias{variance}
\title{Moments of a Distribution}
\usage{
kurtosis_raw(x, ...)

kurtosis_exc(x, ...)

\method{mean}{dst}(x, ...)

\method{median}{dst}(x, ...)

skewness(x, ...)

stdev(x, ...)

\method{stdev}{dst}(x, ...)

variance(x, ...)
}
\arguments{
\item{x}{Object to compute skewness}

\item{...}{Other arguments to pass to specific methods. Currently not used.}
}
\value{
A single numeric.
}
\description{
Get common moment-related quantities of a
distribution: mean, variance, standard deviation (sd),
skewness, and kurtosis.
}
\details{
If there is no method associated with a subclass of
\code{x}, then moments are calculated using \code{stats::integrate()}
from the survival function.
}
\examples{
a <- dst_gpd(0, 1, 0.5)
b <- dst_unif(0, 1)
c <- dst_norm(3, 4)
mean(a)
variance(b)
kurtosis_raw(c)
kurtosis_exc(c)
mean(mix(a, b))
}
