% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_parameters.R
\name{mutate_finite}
\alias{mutate_finite}
\title{Mutate Discrete Values}
\usage{
mutate_finite(distribution, location)
}
\arguments{
\item{distribution}{Finite distribution.}

\item{location}{An expression involving \code{location}.}
}
\value{
The input distribution with discrete values modified according
to the expression in \code{location}.
}
\description{
Transform discrete values in a finite distribution.
}
