% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_unif.R
\name{mutate_parameters}
\alias{mutate_parameters}
\title{Mutate Parameters}
\usage{
mutate_parameters(distribution, ...)
}
\arguments{
\item{distribution}{Parametric distribution}

\item{...}{Named expressions. Names should be parameters names of the
distribution. Expressions can involve computations with other parameters.}
}
\value{
The input distribution, with the parameters modified as specified
in \code{...}.
}
\description{
Modify parameters using tidy evaluation. Lightweight function that only
checks that the parameter you're modifying already exists in the
distribution.
}
\examples{
d <- dst_unif(1, 3)
distplyr:::mutate_parameters(d, min = max - min)
}
