% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{new_distribution}
\alias{new_distribution}
\title{Constructor Function for "dst" Objects}
\usage{
new_distribution(l, variable, ..., class = character())
}
\arguments{
\item{l}{List containing the components of a distribution object.}

\item{variable}{Type of random variable: "continuous", "discrete",
or "mixed".}

\item{...}{Attributes to add to the list.}

\item{class}{If making a subclass, specify its name here.}
}
\description{
Constructor Function for "dst" Objects
}
