% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_finite.R
\name{new_finite}
\alias{new_finite}
\title{Constructor Function for Finite Distributions}
\usage{
new_finite(l, variable, ..., class = character())
}
\arguments{
\item{l}{List containing the components of a step distribution object.}

\item{variable}{Type of random variable: "continuous", "discrete",
or "mixed".}

\item{...}{Attributes to add to the list.}

\item{class}{If making a subclass, specify its name here.}
}
\description{
Constructor Function for Finite Distributions
}
