% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graft_left.R
\name{new_graft}
\alias{new_graft}
\title{Constructor function for graft distributions}
\usage{
new_graft(object, ..., class = character())
}
\arguments{
\item{object}{A special mixture of conditional distributions.}

\item{...}{Attributes to add to the list.}

\item{class}{If making a subclass, specify its name here.}
}
\description{
Constructor function for graft distributions
}
\note{
A graft distribution is a special case of a mixture distribution.
}
