% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-mix.R
\name{new_mix}
\alias{new_mix}
\title{Constructor function for "mix" objects}
\usage{
new_mix(l, variable, ..., class = character())
}
\arguments{
\item{l}{List containing the components of a mixture distribution object.}

\item{variable}{Type of random variable: "continuous", "discrete",
or "mixed".}

\item{...}{Other attributes to add to the list.}

\item{class}{If making a subclass, specify its name here.}
}
\description{
Constructor function for "mix" objects
}
