% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{new_parametric}
\alias{new_parametric}
\title{Constructor Function for "parametric" distribution Objects}
\usage{
new_parametric(l, variable, ..., class = character())
}
\arguments{
\item{l}{List containing the components of a parametric distribution object.}

\item{variable}{Type of random variable: "continuous", "discrete",
or "mixed".}

\item{...}{Attributes to add to the list.}

\item{class}{If making a subclass, specify its name here.}
}
\description{
Constructor Function for "parametric" distribution Objects
}
