% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_dst.R
\name{new_parametric_dst}
\alias{new_parametric_dst}
\title{Constructor Function for "dst" Objects}
\usage{
new_parametric_dst(
  l,
  variable = c("continuous", "discrete", "mixed"),
  class = character()
)
}
\arguments{
\item{l}{List containing the components of a distribution object.}

\item{class}{If making a subclass, specify its name here.}

\item{...}{Attributes to add to the list.}
}
\description{
Constructor Function for "dst" Objects
}
