% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdst.R
\name{new_stepdst}
\alias{new_stepdst}
\title{Constructor Function for Step Distributions}
\usage{
new_stepdst(l, variable, ..., class = character())
}
\arguments{
\item{l}{List containing the components of a step distribution object.}

\item{variable}{Type of random variable: "continuous", "discrete",
or "mixed".}

\item{...}{Attributes to add to the list.}

\item{class}{If making a subclass, specify its name here.}
}
\description{
Constructor Function for Step Distributions
}
