% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_discrete.R
\name{next_discrete}
\alias{next_discrete}
\title{Return neighbouring discrete values from a reference}
\usage{
next_discrete(object, from, n = 1L, ...)
}
\arguments{
\item{object}{Distribution}

\item{from}{Reference value.}

\item{n}{Number of discrete values to find.}
}
\value{
Vector containing the \code{n} discrete points of non-zero
probability using \code{from} as the reference point.
}
\description{
Finds the next \code{n} largest/smallest discrete value in a distribution
from some reference point.
}
\examples{
next_discrete(dst_pois(1), from = 1.3)
prev_discrete(dst_pois(1), from = 3, n = 10)
next_discrete(dst_norm(0, 1), from = 1.3, n = 4)
}
