% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframe.R, R/eval_odds.R
\name{enframe_odds}
\alias{enframe_odds}
\alias{eval_odds}
\title{Odds Function}
\usage{
enframe_odds(..., at, arg_name = ".arg", fn_prefix = "odds", sep = "_")

eval_odds(object, at)
}
\arguments{
\item{at}{Vector of values to evaluate the cdf at.}

\item{arg_name}{Name of the column containing the function arguments, in the
output data frame of \code{enframe_}.}

\item{object}{Distribution object.}
}
\value{
A vector of the evaluated odds, in the case of \code{eval_}; a
data frame or tibble with both the argument and function evaluations,
in the case of \code{enframe_};
or a vectorized function representing the odds function, in the
case of \code{get_}.
}
\description{
Access a distribution's odds function. The odds of an event having
probability \code{p} is \code{p / (1 - p)}.
}
\examples{
d <- dst_pois(1)
eval_odds(d, at = c(1, 2, 2.5))
enframe_odds(d, at = 0:4)
eval_odds(dst_norm(0, 1), at = 0:4, strict = FALSE)
}
\seealso{
Other distributional representations: 
\code{\link{enframe_cdf}()},
\code{\link{enframe_chf}()},
\code{\link{enframe_density}()},
\code{\link{enframe_hazard}()},
\code{\link{enframe_pmf}()},
\code{\link{enframe_quantile}()},
\code{\link{enframe_survival}()}
}
\concept{distributional representations}
