% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_dst.R
\name{parameters}
\alias{parameters}
\title{Parameters of a Parametric Distribution}
\usage{
parameters(object)
}
\arguments{
\item{object}{Distribution object}
}
\description{
Parameters of a Parametric Distribution
}
