% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{is_parametric}
\alias{is_parametric}
\alias{is.parametric}
\title{Parametric Distribution Objects}
\usage{
is_parametric(object)

is.parametric(object)
}
\arguments{
\item{object}{Object to be tested}
}
\description{
Parametric Distribution Objects
}
