% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqdr_funs.R
\name{pdst.dst}
\alias{pdst.dst}
\title{Calculate cdf}
\usage{
\method{pdst}{dst}(object, q)
}
\arguments{
\item{object}{Object of class "dst"}

\item{q}{Vector of values to evaluate the cdf at}
}
\value{
Vector of cdf values if \code{q} is present; otherwise,
returns a function for the cdf.
}
\description{
Calculate cdf
}
\seealso{
\code{\link{qdst}}, \code{\link{ddst}}, \code{\link{rdst}}
}
