% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-stepfun.R
\name{plateaus}
\alias{plateaus}
\title{Extract Heights of a Step Function}
\usage{
plateaus(object)
}
\arguments{
\item{object}{Object of class "stepfun".}
}
\value{
Vector of y values from left to right.
}
\description{
Extracts the heights/"y values"/plateaus of a
step function. Sister function to \code{stats::knots()},
which returns the breakpoints/"x values".
}
