% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst.R
\name{plot.dst}
\alias{plot.dst}
\title{Plot a Distribution}
\usage{
\method{plot}{dst}(
  object,
  what = c("probfn", "cdf", "survival", "quantile", "hazard", "chf"),
  ...
)
}
\arguments{
\item{object}{Distribution object}

\item{what}{Name of the representation to plot.}

\item{...}{Other arguments to pass to the \code{graphics::curve} function.}
}
\description{
Plot a functional representation of a distribution.
Wrapper around the \code{graphics::curve} function.
}
