% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframe.R, R/pmf.R
\name{enframe_pmf}
\alias{enframe_pmf}
\alias{eval_pmf}
\alias{get_pmf}
\title{Probability Mass Function}
\usage{
enframe_pmf(..., at, arg_name = ".arg", fn_name = ".pmf", strict = TRUE)

eval_pmf(object, at, strict = TRUE)

get_pmf(object)
}
\arguments{
\item{at}{Vector of values to evaluate the cdf at.}

\item{arg_name}{Name of the column containing the function arguments, in the
output data frame of \code{enframe_}.}

\item{fn_name}{Name of the column containing the function evaluations, in the
output data frame of \code{enframe_}.}

\item{object}{Distribution object.}
}
\value{
A vector of the evaluated pmf/pdf, in the case of \code{eval_}; a
data frame with both the argument and function evaluations, in the case of
\code{enframe_}; or a vectorized function representing the pmf/pdf, in the
case of \code{get_}.
}
\description{
Access a distribution's probability mass function (pmf).
}
\examples{
d <- dst_empirical(1:10)
eval_pmf(d, at = c(1, 2, 2.5))
enframe_pmf(d, at = 0:4)
pmf <- get_pmf(d)
pmf(0:4)
}
\seealso{
Other distributional representations: 
\code{\link{eval_cdf}()},
\code{\link{eval_density}()},
\code{\link{eval_hazard}()},
\code{\link{eval_quantile}()},
\code{\link{eval_survival}()},
\code{\link{get_chf}()}
}
\concept{distributional representations}
