% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn.dst.R
\name{get_higher}
\alias{get_higher}
\alias{get_lower}
\title{Find an Encapsulating Range of Quantiles}
\usage{
get_higher(object, level)

get_lower(object, level)
}
\arguments{
\item{object}{Distribution}

\item{level}{Quantile level defining the}
}
\description{
Given a quantile level, these functions
will find a quantile that evaluates to
be higher/lower than the provided quantile level.
Intended for internal use only.
}
