% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation-probfn.R
\name{eval_probfn}
\alias{eval_probfn}
\alias{get_probfn}
\alias{enframe_probfn}
\title{Probability density/mass function}
\usage{
eval_probfn(object, at)

get_probfn(object)

enframe_probfn(object, at, arg_name = ".arg", fn_name = ".probfn")
}
\arguments{
\item{object}{Distribution object.}

\item{at}{Vector of values to evaluate the cdf at.}

\item{arg_name}{Name of the column containing the function arguments, in the
output data frame of \code{enframe_}.}

\item{fn_name}{Name of the column containing the function evaluations, in the
output data frame of \code{enframe_}.}
}
\value{
A vector of the evaluated pmf/pdf, in the case of \code{eval_}; a
  data frame with both the argument and function evaluations, in the case of
  \code{enframe_}; or a vectorized function representing the pmf/pdf, in the
  case of \code{get_}.
}
\description{
Access a distribution's probability density or mass function.
}
\examples{
d <- dst_unif(0, 4)
eval_probfn(d, at = 0:4)
enframe_probfn(d, at = 0:4)
probfn <- get_probfn(d)
probfn(0:4)
}
\seealso{
Other distributional representations: 
\code{\link{eval_cdf}()},
\code{\link{eval_hazard}()},
\code{\link{eval_quantile}()},
\code{\link{eval_survival}()},
\code{\link{get_chf}()}
}
\concept{distributional representations}
