% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqdr_funs.R
\name{qdst}
\alias{qdst}
\title{Calculate quantile function}
\usage{
qdst(object, p)
}
\arguments{
\item{object}{Object of class "dst"}

\item{p}{Vector of values to evaluate the quantile function at}
}
\value{
Vector of quantiles if \code{p} is present; otherwise,
returns a function for the quantile function.
}
\description{
Calculate quantile function
}
\seealso{
\code{\link{pdst}}, \code{\link{ddst}}, \code{\link{rdst}}
}
