% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframe.R, R/eval_quantile.R
\name{enframe_quantile}
\alias{enframe_quantile}
\alias{eval_quantile}
\alias{get_quantile}
\alias{get_quantile.dst}
\alias{eval_quantile.dst}
\title{Distribution Quantiles}
\usage{
enframe_quantile(..., at, arg_name = ".arg", fn_prefix = "quantile", sep = "_")

eval_quantile(object, at, ...)

get_quantile(object, ...)

\method{get_quantile}{dst}(object, tol = 1e-06, maxiter = 1000, ...)

\method{eval_quantile}{dst}(object, at, tol = 1e-06, maxiter = 1000, ...)
}
\arguments{
\item{...}{Other arguments to pass to specific methods.}

\item{at}{Vector of values to evaluate the cdf at.}

\item{arg_name}{Name of the column containing the function arguments, in the
output data frame of \code{enframe_}.}

\item{object}{Distribution object.}

\item{tol}{Error tolerance when using an algorithm to find the left-inverse
of the cdf.}

\item{maxiter}{Maximum number of iterations when using an algorithm to find
the left-inverse of the cdf.}
}
\value{
A vector of the evaluated quantiles, in the case of
\code{eval_}; a data frame with both the argument and
function evaluations, in the case of \code{enframe_};
or a vectorized function representing the quantile function, in the
case of \code{get_}.
}
\description{
Access a distribution's quantiles.
}
\examples{
d <- dst_unif(0, 4)
eval_quantile(d, at = 0:4)
enframe_quantile(d, at = 0:4)
qf <- get_quantile(d)
qf(0:4)
}
\seealso{
Other distributional representations: 
\code{\link{enframe_cdf}()},
\code{\link{enframe_chf}()},
\code{\link{enframe_density}()},
\code{\link{enframe_hazard}()},
\code{\link{enframe_odds}()},
\code{\link{enframe_pmf}()},
\code{\link{enframe_survival}()}
}
\concept{distributional representations}
