% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities_list.R
\docType{data}
\name{quantities}
\alias{quantities}
\title{Specific Formulas for Quantities}
\format{
An object of class \code{list} of length 5.
}
\usage{
quantities
}
\description{
Formulas for quantities (such as mean, variance, skewness, EVI, etc.)
of select parametric distributions.
}
\details{
A list, where each distribution gets a (named) entry, with name
given by the suffix of \code{dst_} (such as "norm", "unif", etc.). Each
distribution's entry is itself a named list of expressions, where the
name is the name of the quantity matching the distplyr function name
(such as "mean", "kurtosis_exc", etc.). Each expression is allowed to
refer to the distribution's parameters by name.
}
\note{
Although R allows us to evaluate distributional representations
of certain parametric distributions through functions with
\code{p}, \code{d}, \code{q}, and \code{r} prefixes (such as \code{pnorm()}, \code{dnorm()}, etc.),
R does not "come with" formulas for quantities such as mean, variance,
EVI, etc. Although these quantities can be computed from a distributional
representation (such as integrating the quantile function to get the mean),
it's often inefficient to rely on such computations. We therefore include
formulas here, and check them using testthat.
}
\keyword{datasets}
