% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities-mix.R
\name{range.mix}
\alias{range.mix}
\title{#' @export
evi.mix <- function(distribution, ...) {
with(distribution[\link{"components"}], {
right_ends <- vapply(
distributions, function(d) distionary::range(d)\link{2L},
FUN.VALUE = numeric(1L)
)
max_end <- max(right_ends)
has_max_ends <- right_ends == max_end
evis <- vapply(distributions, distionary::evi, FUN.VALUE = numeric(1L))
final_sign <- if (max_end < Inf) -1 else 1
final_sign * max(abs(evis\link{has_max_ends}))
})
}
Range of a mixture distribution}
\usage{
\method{range}{mix}(distribution, ...)
}
\arguments{
\item{distribution}{Distribution object.}

\item{...}{Not used.}
}
\value{
Vector of length two, indicating the support of the distribution.
}
\description{
#' @export
evi.mix <- function(distribution, ...) {
with(distribution[\link{"components"}], {
right_ends <- vapply(
distributions, function(d) distionary::range(d)\link{2L},
FUN.VALUE = numeric(1L)
)
max_end <- max(right_ends)
has_max_ends <- right_ends == max_end
evis <- vapply(distributions, distionary::evi, FUN.VALUE = numeric(1L))
final_sign <- if (max_end < Inf) -1 else 1
final_sign * max(abs(evis\link{has_max_ends}))
})
}
Range of a mixture distribution
}
