% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqdr_funs.R
\name{rdst.dst}
\alias{rdst.dst}
\title{Generate random sample from a distribution}
\usage{
\method{rdst}{dst}(object, n)
}
\arguments{
\item{object}{Object of class "dst"}

\item{n}{Number of observations to generate}
}
\value{
Vector of independent values drawn from the inputted distribution
if \code{n} is present; otherwise,
returns a function for a random number generator.
}
\description{
Generate random sample from a distribution
}
\seealso{
\code{\link{qdst}}, \code{\link{ddst}}, \code{\link{pdst}}
}
