% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realise.R
\name{realise}
\alias{realise}
\alias{realize}
\alias{observe}
\title{Generate a Sample from a Distribution}
\usage{
realise(distribution, n = 1, ...)

realize(distribution, n = 1)

observe(distribution, n = 1)
}
\arguments{
\item{distribution}{Distribution.}

\item{n}{Number of observations to generate}

\item{...}{Other arguments to pass to \code{\link{eval_quantile}}}
}
\value{
Vector of independent values drawn from the inputted distribution
}
\description{
Draw \code{n} independent observations from a distribution.
}
\note{
\code{realise()}, \code{realize()}, and \code{observe()} all do the same thing.
}
