% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_unif.R
\name{resolve_if_possible}
\alias{resolve_if_possible}
\title{Resolve an input if possible}
\usage{
resolve_if_possible(x)
}
\arguments{
\item{x}{A line of code.}
}
\value{
A list: the \code{$outcome} entry contains the
evaluated input if evaluation is possible, or a quosure
of the input if not possible; \code{$resolved} is a
logical entry indicating whether or not the input was
able to be resolved (\code{TRUE} if so, \code{FALSE} if not).
}
\description{
This function tries to evaluate its argument, providing
the outcome if succeeding, and a quosure if failing.
}
\examples{
resolve_if_possible(cowabunga)
resolve_if_possible(42)
cowabunga <- 42
resolve_if_possible(cowabunga)
}
