% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substail.R
\name{right_connect}
\alias{right_connect}
\title{Replace the tail of a distribution}
\usage{
right_connect(left_dst, right_dst, sep_x)
}
\arguments{
\item{left_dst, right_dst}{Distributions to connect}

\item{sep_x}{Value on the domain of the cdf to connect at.}
}
\value{
Object of class "dst"
}
\description{
\code{right_connect()} keeps the left cdf unchanged to the left of
sep_val, and makes a continuous connection with the right cdf
(rescaled as appropriate).
\code{left_connect()} keeps the right cdf unchanged to the right of
sep_val, and makes a continuous connection with the left cdf
(rescaled as appropriate),
}
