% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{rv_locscale}
\alias{rv_locscale}
\title{Linearly Transform a Distribution}
\usage{
rv_locscale(.dst, loc, scale)
}
\arguments{
\item{.dst}{Original distribution to transform}

\item{loc}{Single numeric.}

\item{scale}{Single non-negative numeric.}
}
\value{
Object of class "dst" of the transformed random variable.
}
\description{
If \code{X} has distribution \code{.dst}, then this function
returns the distribution of \code{scale*X + loc}. A wrapper
around \code{\link{rv_transform}}.
}
\seealso{
\code{\link{rv_transform}}
}
