% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{rv_transform}
\alias{rv_transform}
\title{Transform a Distribution}
\usage{
rv_transform(.dst, g, ginv, gprime, ginvprime)
}
\arguments{
\item{.dst}{Original distribution to transform}

\item{g}{A (strictly increasing) function, vectorized.}

\item{ginv}{Inverse function of \code{g}, vectorized.}

\item{gprime, ginvprime}{Derivative functions of \code{g} and \code{ginv},
vectorized.
Not required if there is no density, and if there is a density, only
one is required.}
}
\value{
Object of class "dst" of the transformed random variable.
}
\description{
If \code{X} has distribution \code{.dst}, then this function
returns the distribution of \code{g(X)}
}
\details{
Functions should apply to all real numbers.
}
