% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_hazard.R
\name{scale_hazard}
\alias{scale_hazard}
\title{Scale a distribution's hazard function}
\usage{
scale_hazard(distribution, by)
}
\arguments{
\item{distribution}{A probability distribution.}

\item{by}{Scaling factor; a single positive number.}
}
\value{
A probability distribution whose hazard function is the
original \code{distribution}'s hazard function with ordinates multiplied by \code{by}.
}
\description{
Makes a new distribution whose hazard function (ordinates) are a
multiple of a baseline distribution's hazard function. This is
particularly useful for proportional hazards models such as the Cox
model.
}
\note{
Currently this function only applies to continuous distributions,
as it's unclear whether an equivalent would make sense otherwise.
}
