% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{sd.default}
\alias{sd.default}
\title{Dispatch sd function from stats}
\usage{
\method{sd}{default}(x, ...)
}
\arguments{
\item{x}{Object for which to calculate standard deviation}

\item{...}{Other arguments to pass to \code{sd} function
from the stats package.}
}
\value{

}
\description{
Dispatch sd function from stats
}
