% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_representation.R
\name{set_cdf}
\alias{set_cdf}
\alias{set_survival}
\alias{set_quantile}
\alias{set_density}
\alias{set_pmf}
\alias{set_hazard}
\alias{set_chf}
\title{Set a representation of a distribution}
\usage{
set_cdf(object, fn)

set_survival(object, fn)

set_quantile(object, fn)

set_density(object, fn)

set_pmf(object, fn)

set_hazard(object, fn)

set_chf(object, fn)
}
\arguments{
\item{object}{Distribution object.}

\item{fn}{Vectorized function of the representation.}
}
\value{
A distribution object, as given in \code{object},
containing the specified representation.
}
\description{
Manually specify a representation of a distribution.
}
