% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifications-dst_gpd.R
\name{shift.gpd}
\alias{shift.gpd}
\title{Linear Transformations}
\usage{
shift.gpd(distribution, constant)
}
\arguments{
\item{distribution}{A probability distribution.}

\item{constant}{A single numeric by which to shift or multiply the
distribution by.}
}
\value{
A distribution, shifted or multiplied by the constant.
Specifically, a distribution with subclass "shift", "scale", or "flip".
}
\description{
\code{shift()} a distribution by adding a constant, or \code{multiply()} a
distribution by a constant. \code{flip()} is a specific case of multiplying
a distribution by \code{-1}, resulting in "flipping" the distribution about 0.
}
\details{
Specifically, if \code{X} is a random variable coming from a
distribution, then the resulting distributions are as follows:
\itemize{
\item For \code{shift()}, is the distribution of \code{X + constant}.
\item For \code{multiply()}, is the distribution of \code{X * constant}.
\item For \code{flip()}, is the distribution of \code{-X}.
}

Although the \code{multiply()} function accepts negative constants,
the corresponding "scale" distribution class only holds positive
constants, delegating a potential negative sign to the "flip" class.
}
\examples{
d_pois <- distionary::dst_pois(1.1)
d_norm <- distionary::dst_norm(4, 1)
d_unif <- distionary::dst_unif(0, 1)

# Shift a Poisson distribution by 1.
shift(d_pois, 1)
d_pois + 1

# Multiply a Uniform distribution by 2.
multiply(d_unif, 2)
d_unif * 2

# Flip a Normal distribution.
flip(d_norm)
-d_norm

# Combine multiple operations:
4 - 2 * d_pois
}
\seealso{
\code{invert()}
}
