% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_left.R, R/slice_right.R
\name{slice_left}
\alias{slice_left}
\alias{slice_right}
\title{Conditional Distributions}
\usage{
slice_left(distribution, breakpoint, include = TRUE, ...)

slice_right(distribution, breakpoint, include = TRUE, ...)
}
\arguments{
\item{distribution}{Distribution to slice.}

\item{breakpoint}{Point at which to slice (single numeric).}

\item{include}{Logical; should the breakpoint be removed as well?
This is only realistically relevant if the
breakpoint has a non-zero probability of occurrence.}

\item{...}{Other arguments to pass to specific methods. Currently unused.}
}
\value{
A conditional distribution.
}
\description{
\code{slice_left()} removes probability to the left of some breakpoint,
conditioning the random variable to be bigger than the breakpoint.
\code{slice_right()} does the opposite: removes probability to the right,
conditioning to be smaller than the breakpoint.
}
\examples{
library(magrittr)
dst_norm(0, 1) \%>\%
  slice_left(-2) \%>\%
  slice_right(2) \%>\%
  enframe_cdf(at = -3:3)

d <- dst_empirical(c(2, 5, 6, 9, 11))
d \%>\%
  slice_left(5) \%>\%
  eval_pmf(at = 5)
d \%>\%
  slice_left(5, include = TRUE) \%>\%
  eval_pmf(at = 5)
}
