% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdev.R
\name{stdev.default}
\alias{stdev.default}
\title{Dispatch sd function from stats}
\usage{
\method{stdev}{default}(x, ...)
}
\arguments{
\item{x}{Object for which to calculate standard deviation}

\item{...}{Other arguments to pass to \code{sd} function
from the stats package.}
}
\description{
Dispatch sd function from stats
}
