% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdst.R
\name{stepdst}
\alias{stepdst}
\alias{is_stepdst}
\alias{is.stepdst}
\title{Create a Step Distribution}
\usage{
stepdst(y, data, weights = 1, ...)

is_stepdst(x)

is.stepdst(x)
}
\arguments{
\item{y}{Outcomes to comprise the distribution. Should either
evaluate to a vector, or be a name in the specified data.}

\item{data}{Data frame, list, or environment
containing the outcome name in \code{y}. If missing,
\code{y} will be evaluated in the parent frame.}

\item{weights}{Weights corresponding to the outcomes in \code{y}.
Must not be negative, but need not sum to 1. If \code{data}
is provided, the data will be searched for the name provided in
this argument.}

\item{...}{Additional arguments to be passed to \code{\link{dst}}.}
}
\value{
A "stepdst" object, which is also a "dst" object,
containing a cdf, quantile function, and random number generator.
The cdf is a right-continuous step function, and the quantile function is
a left-continuous step function.
If you'd like to add other
functions through the \code{\link{dst}} function, you can do so
via \code{...}.
}
\description{
A step distribution is one where the cdf
and quantile function are step functions.
This includes empirical distributions.
\code{stepdst()} facilitates the creation of
such a distribution by specifying the
observations/breakpoints, along with their weights.
}
