% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdst.R
\name{steps}
\alias{steps}
\title{Get Step Points from a Step Distribution}
\usage{
steps(object)
}
\arguments{
\item{object}{A step distribution object.}
}
\value{
Data frame with two columns: column
\code{y} contains the outcomes corresponding to
the step discontinuities of the cdf, and column
\code{tau} contains the non-zero plateaus of the
cdf (so that the cdf at \code{y} evaluates to
\code{tau})
}
\description{
Step points are the coordinates marking the
positions of the step discontinuities in a step
distribution. They are the left-most points of
each non-zero plateau of the cdf.
}
