% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substail.R
\name{substail_old}
\alias{substail_old}
\title{Substitute a new distribution for the tail}
\usage{
substail_old(pdist1, pdist2, qdist1, qdist2, sep_x)
}
\arguments{
\item{pdist1, pdist2, qdist1, qdist2}{Vectorized functions representing the lower (1)
and upper (2) distribution functions (pdist) and quantile functions (qdist).}

\item{sep_x}{Single numeric representing the value of the random
variable defining the separation point.}
}
\value{
Named list of two components: pdist and qdist, which are vectorized
functions that preserve NA's.
}
\description{
Constructs a pdist that is the original pdist up to some cutpoint (on the
x-axis), then continues with pdist2.
}
\details{
Gives the following distribution: if W1 and W2 have
pdist's pdist1 and pdist2, let p = pdist1(sep_x), and let L be Bernoulli(p),
whose outcome indicates whether the lower distribution will be used or not.
Then the resulting pdist is that of W1 if L=1, and W2 if L=0.
}
