% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframe.R, R/eval_survival.R
\name{enframe_survival}
\alias{enframe_survival}
\alias{eval_survival}
\alias{get_survival}
\title{Survival Function}
\usage{
enframe_survival(..., at, arg_name = ".arg", fn_prefix = "survival", sep = "_")

eval_survival(object, at)

get_survival(object)
}
\arguments{
\item{at}{Vector of values to evaluate the cdf at.}

\item{arg_name}{Name of the column containing the function arguments, in the
output data frame of \code{enframe_}.}

\item{object}{Distribution object.}
}
\value{
A vector of the evaluated survival function, in the case of
\code{eval_}; a data frame with both the argument and function evaluations,
in the case of \code{enframe_}; or a vectorized function representing the
survival function, in the case of \code{get_}.
}
\description{
Access a distribution's survival function.
}
\examples{
d <- dst_unif(0, 4)
eval_survival(d, at = 0:4)
enframe_survival(d, at = 0:4)
survival <- get_survival(d)
survival(0:4)
}
\seealso{
Other distributional representations: 
\code{\link{enframe_cdf}()},
\code{\link{enframe_chf}()},
\code{\link{enframe_density}()},
\code{\link{enframe_hazard}()},
\code{\link{enframe_odds}()},
\code{\link{enframe_pmf}()},
\code{\link{enframe_quantile}()}
}
\concept{distributional representations}
