% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-stepfun.R
\name{swap_step_continuity_direction}
\alias{swap_step_continuity_direction}
\title{Swap Direction of Continuity of a Step Function}
\usage{
swap_step_continuity_direction(stepfun)
}
\arguments{
\item{stepfun}{Object of class "stepfun".}
}
\value{
Another stepfun with the continuity swapped
between left and right continuous. That is, the only
way the new function evaluates differently is at the
breakpoints.
}
\description{
Swap Direction of Continuity of a Step Function
}
