% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_dst.R
\name{tbl_dst}
\alias{tbl_dst}
\title{Create a tbl_dst}
\usage{
tbl_dst(.data, ...)
}
\arguments{
\item{.data}{Data frame or tibble}

\item{...}{variables to connect a distribution to, some perhaps set
to a \code{dst} object. Tidyselect should work, too.}
}
\value{
The original .data object, with a new "distributions"
attribute: a list whose names match the names of .data, and whose
entries are either a distribution (if that variable was specified),
or NA if not. Variables not set to a specific distribution will have
an empirical distribution.
}
\description{
Connect a distribution to variables in a data frame.
Experimental and not robust to downstream manipulation.
The handling of such objects probably
deserves its own package.
}
\details{
Distributions are set in the order that they show up in \code{...}.
}
\note{
No precaution has yet to be taken as to how the \code{distributions}
attribute is handled downstream. \code{mutate()} etc. at your own risk.
}
