% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation-cdf.R
\name{eval_cdf}
\alias{eval_cdf}
\alias{get_cdf}
\title{Distribution Function}
\usage{
eval_cdf(object, at)

get_cdf(object)
}
\arguments{
\item{object}{Object of class "dst"}

\item{at}{Vector of values to evaluate the cdf at}
}
\value{
Vector of cdf values
}
\description{
Get or evaluate a distribution's cdf
(Cumulative Distribution Function).
}
\examples{
d <- dst_unif(0, 4)
eval_cdf(d, at = 0:4)
cdf <- get_cdf(d)
cdf(0:4)
}
\seealso{
\code{\link{eval_quantile}},
\code{\link{eval_probfn}},
\code{\link{realise}},
\code{\link{eval_hazard}},
\code{\link{eval_survival}}
}
