% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframe.R, R/eval_cdf.R
\name{enframe_cdf}
\alias{enframe_cdf}
\alias{eval_cdf}
\alias{get_cdf}
\title{(Cumulative) Distribution Function}
\usage{
enframe_cdf(..., at, arg_name = ".arg", fn_prefix = "cdf", sep = "_")

eval_cdf(object, at)

get_cdf(object)
}
\arguments{
\item{at}{Vector of values to evaluate the cdf at.}

\item{arg_name}{Name of the column containing the function arguments, in the
output data frame of \code{enframe_}.}

\item{object}{Distribution object.}

\item{fn_name}{Name of the column containing the function evaluations, in the
output data frame of \code{enframe_}.}
}
\value{
A vector of the evaluated cdf, in the case of \code{eval_}; a data
frame with both the argument and function evaluations, in the case of
\code{enframe_}; or a vectorized function representing the cdf, in the case
of \code{get_}.
}
\description{
Access a distribution's cumulative distribution function (cdf).
}
\examples{
d <- dst_unif(0, 4)
eval_cdf(d, at = 0:4)
enframe_cdf(d, at = 0:4)
cdf <- get_cdf(d)
cdf(0:4)
}
\seealso{
Other distributional representations: 
\code{\link{enframe_chf}()},
\code{\link{enframe_density}()},
\code{\link{enframe_hazard}()},
\code{\link{enframe_odds}()},
\code{\link{enframe_pmf}()},
\code{\link{enframe_quantile}()},
\code{\link{enframe_survival}()}
}
\concept{distributional representations}
