% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_cdf.R
\name{eval_cdf}
\alias{eval_cdf}
\alias{enframe_cdf}
\title{Cumulative Distribution Function}
\usage{
eval_cdf(distribution, at)

enframe_cdf(..., at, arg_name = ".arg", fn_prefix = "cdf", sep = "_")
}
\arguments{
\item{distribution, ...}{A distribution, or possibly multiple
distributions in the case of \code{...}.}

\item{at}{Vector of values to evaluate the cdf at. Must be named when using
in \code{enframe_}.}

\item{arg_name}{For \code{enframe_}, name of the column containing
the function arguments.}

\item{fn_prefix}{For \code{enframe_}, name of the function to
appear in the column(s).}

\item{sep}{When \code{enframe}'ing more than one distribution, the
character that will be separating the \code{fn_name} and the distribution name.}
}
\value{
The evaluated cdf in vector form (for \code{eval_}) and data frame
or tibble form (for \code{enframe_}).
}
\description{
Access a distribution's cumulative distribution function (cdf).
}
\examples{
d <- dst_unif(0, 4)
eval_cdf(d, at = 0:4)
enframe_cdf(d, at = 0:4)
enframe_cdf(d, d + 1, at = 0:4)
}
\seealso{
Other distributional representations: 
\code{\link{eval_chf}()},
\code{\link{eval_density}()},
\code{\link{eval_hazard}()},
\code{\link{eval_odds}()},
\code{\link{eval_pmf}()},
\code{\link{eval_quantile}()},
\code{\link{eval_survival}()}
}
\concept{distributional representations}
