% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation-chf.R
\name{get_chf}
\alias{get_chf}
\alias{eval_chf}
\alias{enframe_chf}
\title{Cumulative Hazard Function}
\usage{
get_chf(object)

eval_chf(object, at)

enframe_chf(object, at, arg_name = ".arg", fn_name = ".chf")
}
\arguments{
\item{object}{Object of class "dst"}

\item{at}{Vector of values to evaluate the cumulative hazard at.}
}
\value{
Vector of the evaluated hazard function
}
\description{
Cumulative Hazard Function
}
\seealso{
\code{\link{eval_quantile}},
\code{\link{eval_probfn}},
\code{\link{eval_cdf}},
\code{\link{eval_survival}}
}
