% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_degenerate.R
\name{dst_degenerate}
\alias{dst_degenerate}
\alias{is_degenerate}
\alias{is.degenerate}
\title{Make a Degenerate Distribution}
\usage{
dst_degenerate(location)

is_degenerate(object)

is.degenerate(object)
}
\arguments{
\item{location}{Parameter of the distribution family.}

\item{object}{Object to test}
}
\value{
Object of class "dst".
}
\description{
Makes a distribution belonging to the degenerate family of
distributions. That is, distributions of fixed values.
}
\examples{
require(graphics)
d <- dst_degenerate(5)
plot(d, "quantile")
}
