% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{eval_density}
\alias{eval_density}
\alias{get_density}
\alias{enframe_density}
\title{Probability Density Function}
\usage{
eval_density(object, at, strict = TRUE)

get_density(object)

enframe_density(object, at, arg_name = ".arg", fn_name = ".density")
}
\arguments{
\item{object}{Distribution object.}

\item{at}{Vector of values to evaluate the cdf at.}

\item{arg_name}{Name of the column containing the function arguments, in the
output data frame of \code{enframe_}.}

\item{fn_name}{Name of the column containing the function evaluations, in the
output data frame of \code{enframe_}.}
}
\value{
A vector of the evaluated density, in the case of \code{eval_}; a
data frame with both the argument and function evaluations, in the case of
\code{enframe_}; or a vectorized function representing the density, in the
case of \code{get_}.
}
\description{
Access a distribution's probability density function (pdf).
}
\examples{
d <- dst_unif(0, 4)
eval_density(d, at = 0:4)
enframe_density(d, at = 0:4)
density <- get_density(d)
density(0:4)
}
\seealso{
Other distributional representations: 
\code{\link{eval_cdf}()},
\code{\link{eval_hazard}()},
\code{\link{eval_pmf}()},
\code{\link{eval_quantile}()},
\code{\link{eval_survival}()},
\code{\link{get_chf}()}
}
\concept{distributional representations}
