% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_density.R
\name{eval_density}
\alias{eval_density}
\alias{enframe_density}
\title{Probability Density Function}
\usage{
eval_density(distribution, at, strict = TRUE)

enframe_density(
  ...,
  at,
  arg_name = ".arg",
  fn_prefix = "density",
  sep = "_",
  strict = TRUE
)
}
\arguments{
\item{distribution}{A distribution, or possibly multiple
distributions in the case of \code{...}.}

\item{at}{Vector of values to evaluate the cdf at. Must be named when using
in \code{enframe_}.}

\item{strict}{Only evaluate when the density exists? \code{TRUE} if so;
if \code{FALSE}, evaluates the derivative of the cdf.}

\item{...}{A distribution, or possibly multiple
distributions in the case of \code{...}.}

\item{arg_name}{For \code{enframe_}, name of the column containing
the function arguments.}

\item{fn_prefix}{For \code{enframe_}, name of the function to
appear in the column(s).}

\item{sep}{When \code{enframe}'ing more than one distribution, the
character that will be separating the \code{fn_name} and the distribution name.}
}
\value{
The evaluated density in vector form (for \code{eval_}) and data frame
or tibble form (for \code{enframe_}).
}
\description{
Access a distribution's probability density function (pdf).
}
\examples{
d <- dst_unif(0, 4)
eval_density(d, at = 0:4)
enframe_density(d, d + 1, at = 0:4)
eval_density(dst_pois(1), at = 0:4, strict = FALSE)
}
\seealso{
Other distributional representations: 
\code{\link{eval_cdf}()},
\code{\link{eval_chf}()},
\code{\link{eval_hazard}()},
\code{\link{eval_odds}()},
\code{\link{eval_pmf}()},
\code{\link{eval_quantile}()},
\code{\link{eval_survival}()}
}
\concept{distributional representations}
