% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discontinuities.R, R/dst_degenerate.R,
%   R/dst_finite.R, R/dst_gpd.R, R/dst_lnorm.R, R/dst_norm.R, R/dst_pois.R,
%   R/dst_unif.R, R/mix.R
\name{discontinuities}
\alias{discontinuities}
\alias{aggregate_weights}
\alias{discontinuities.degenerate}
\alias{discontinuities.finite}
\alias{discontinuities.gpd}
\alias{discontinuities.lnorm}
\alias{discontinuities.norm}
\alias{discontinuities.pois}
\alias{discontinuities.unif}
\alias{discontinuities.mix}
\title{Discontinuities in a Distribution}
\usage{
discontinuities(object, from, to, ...)

aggregate_weights(y, weights, sum_to_one = FALSE)

\method{discontinuities}{degenerate}(object, from = -Inf, to = Inf, ...)

\method{discontinuities}{finite}(object, from = -Inf, to = Inf, ...)

\method{discontinuities}{gpd}(object, from = -Inf, to = Inf, ...)

\method{discontinuities}{lnorm}(object, from = -Inf, to = Inf, ...)

\method{discontinuities}{norm}(object, from = -Inf, to = Inf, ...)

\method{discontinuities}{pois}(object, from = -Inf, to = Inf, ...)

\method{discontinuities}{unif}(object, from = -Inf, to = Inf, ...)

\method{discontinuities}{mix}(object, from = -Inf, to = Inf, ...)
}
\arguments{
\item{object}{A distribution object.}

\item{y}{Vector of outcomes.}

\item{weights}{Vector of weights, one for each of \code{y}.
These need not sum to one, but must not be negative.}

\item{sum_to_one}{Logical; should the weights be normalized
to sum to 1? Default is FALSE.}
}
\value{
Data frame with the following columns:
\enumerate{
\item \code{location}: Increasing vector of unique values of \code{y}
having positive weight.
\item \code{size}: Weights corresponding to each outcome.
}
}
\description{
Discontinuity points are the coordinates marking the
positions of step discontinuities in a
distribution. They are the upper points at
each discontinuity in the cdf, together
with the size of the discontinuity gap.
}
\details{
For a vector of outcomes \code{y} with a
matching vector of \code{weights},
\code{aggregate_weights()} provides a single non-zero, non-NA
weight per unique value of \code{y}.
}
\examples{
a <- dst_empirical(1:5)
discontinuities(a)
b <- dst_norm(0, 1)
discontinuities(b)
c <- graft_right(a, b, sep_y = 3.5)
discontinuities(c)
}
