% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretes.R
\name{next_discrete_finite}
\alias{next_discrete_finite}
\alias{prev_discrete_finite}
\alias{next_discrete_natural}
\alias{prev_discrete_natural}
\title{Helper functions for finding discrete values}
\usage{
next_discrete_finite(x, from, n, include_from)

prev_discrete_finite(x, from, n, include_from)

next_discrete_natural(from, n, include_from)

prev_discrete_natural(from, n, include_from)
}
\arguments{
\item{x}{A vector of all possible discrete values.}

\item{from}{Reference value.}

\item{n}{Number of discrete values to find.}

\item{include_from}{Logical; should the \code{from} value be included
in the query? Should the \code{to} value?}
}
\description{
\verb{*_discrete_finite} finds discrete values from a
finite set of possibilities. \verb{*_discret_natural} finds
discrete values amongst the natural numbers (including 0).
}
