% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distplyr.R
\docType{package}
\name{distplyr-package}
\alias{distplyr}
\alias{distplyr-package}
\title{distplyr: Manipulation of Probability Distributions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Manipulating probability distributions, using verbs such as `mix`, `shift`, `maximize`, `invert`, and others. This package builds on the distionary package and belongs in the wider probaverse suite of R packages to support statistical modeling and simulations.
}
\details{
The distplyr package provides tools for manipulating probability distributions
using intuitive syntax. Key features include:

\strong{Arithmetic operators}: Use \code{+}, \code{-}, \code{*}, \code{/}, and \code{^} with distributions.
See \code{\link[=Ops.dst]{Ops.dst()}} for details.

\strong{Mathematical functions}: Use \code{log()}, \code{exp()}, \code{log10()}, and \code{sqrt()} with
distributions. See \code{\link[=Math.dst]{Math.dst()}} for details.

\strong{Transformation functions}: Use \code{\link[=shift]{shift()}}, \code{\link[=multiply]{multiply()}}, \code{\link[=flip]{flip()}}, and \code{\link[=invert]{invert()}}
to transform distributions. See \code{?linear_transform} for details.

\strong{Mixing distributions}: Use \code{\link[=mix]{mix()}} to create mixture distributions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://distplyr.probaverse.com/}
}

}
\author{
\strong{Maintainer}: Vincenzo Coia \email{vincenzo.coia@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item Amogh Joshi [contributor]
  \item Shuyi Tan [contributor]
  \item Zhipeng Zhu [contributor]
}

}
