% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distplyr.R
\docType{package}
\name{distplyr-package}
\alias{distplyr}
\alias{distplyr-package}
\title{distplyr: Manipulation of Univariate Distributions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides an interface for manipulating univariate probability distributions, using verbs such as `graft`, `slice`, `mix`, `shift`, and others.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://distplyr.probaverse.com/}
}

}
\author{
\strong{Maintainer}: Vincenzo Coia \email{vincenzo.coia@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item Amogh Joshi [contributor]
  \item Shuyi Tan [contributor]
  \item Zhipeng Zhu [contributor]
}

}
