% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distplyr.r
\docType{package}
\name{distplyr-package}
\alias{distplyr}
\alias{distplyr-package}
\title{distplyr: Manipulation of Univariate Distributions}
\description{
Provides an interface for manipulating univariate probability distributions, using verbs such as `graft`, `slice`, `mix`, `shift`, and others.
}
\author{
\strong{Maintainer}: Vincenzo Coia \email{vincenzo.coia@stat.ubc.ca}

Authors:
\itemize{
  \item Amogh Joshi
  \item Shuyi Tan
  \item Zhipeng Zhu
}

}
