% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{is_distribution}
\alias{is_distribution}
\alias{is.distribution}
\alias{distribution}
\title{Distribution Objects}
\usage{
is_distribution(object)

is.distribution(object)

distribution(variable = c("continuous", "discrete", "mixed"))
}
\arguments{
\item{object}{Object to be tested}

\item{variable}{Is this variable continuous, discrete, or mixed?}
}
\value{
A distribution object with nothing in it.
}
\description{
Currently, this function makes a distribution object with nothing in it. The
idea is that you can then set things downstream, with functions such as
set_cdf() and set_mean(). The idea behind this function is expected to be in
flux.
}
