% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst.R
\name{dst}
\alias{dst}
\title{Make a Distribution}
\usage{
dst(fun_cumu, fun_quant, fun_prob, fun_rand, fun_surv, name = NULL,
  param = NULL, prop = NULL)
}
\arguments{
\item{fun_cumu, fun_quant}{cdf and quantile function for a distribution,
respectively. Must be supplied.}

\item{fun_prob, fun_rand}{probability mass/density function and
random number generator for a distribution (a function of n), respectively.
Optional. \code{fun_rand} will be calculated from \code{fun_quant}
if missing.}

\item{name}{A name for the distribution (such as a parametric family name)}

\item{param}{Parameters for the distribution, if parameteric.}

\item{prop}{Properties of the distribution, such as mean, variance,
EVI, etc. (of your choosing).}
}
\value{
An object of class "dst", which (for now) is a list holding
the above arguments, including survival and hazard functions.
}
\description{
Make a distribution object.
}
