% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst.R
\name{dst}
\alias{dst}
\alias{is_dst}
\alias{is.dst}
\title{Make a blank distribution}
\usage{
dst(variable = c("continuous", "discrete", "mixed"))

is_dst(object)

is.dst(object)
}
\arguments{
\item{variable}{Is this variable continuous, discrete, or mixed?}

\item{object}{Object to be tested}
}
\value{
A distribution object with nothing in it.
}
\description{
Currently, this function makes a distribution object with nothing in it. The
idea is that you can then set things downstream, with functions such as
set_cdf() and set_mean(). The idea behind this function is expected to be in
flux.
}
