% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_unif.R
\name{dst_unif}
\alias{dst_unif}
\title{Make a Uniform distribution}
\usage{
dst_unif(min = 0, max = 1)
}
\arguments{
\item{min, max}{Parameters of the distribution family.}
}
\value{
Object of class "dst".
}
\description{
Makes a distribution belonging to the continuous uniform family of
distributions.
}
