% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_unif.R
\name{dst_unif}
\alias{dst_unif}
\title{Uniform Distribution}
\usage{
dst_unif(min, max)
}
\arguments{
\item{min, max}{Minimum and maximum of the distribution.}
}
\description{
Makes a distribution belonging to the family of
Uniform distributions.
}
\examples{
dst_unif(0, 1)
}
