% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_fn.R
\name{eval_randfn}
\alias{eval_randfn}
\title{Generate random sample from a distribution}
\usage{
eval_randfn(object, at)
}
\arguments{
\item{object}{Object of class "dst"}

\item{at}{Number of observations to generate}
}
\value{
Vector of independent values drawn from the inputted distribution
}
\description{
Generate random sample from a distribution
}
\seealso{
\code{\link{eval_quantfn}},
\code{\link{eval_probfn}},
\code{\link{eval_cdf}},
\code{\link{eval_hazfn}},
\code{\link{eval_surv}}
}
