% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evi.R
\name{evi.dst}
\alias{evi.dst}
\title{Extreme Value Index}
\usage{
\method{evi}{dst}(x, ...)
}
\arguments{
\item{x}{Distribution object to obtain EVI from.}

\item{...}{Not used.}
}
\value{
A single numeric.
}
\description{
Extreme Value Index
}
\details{
Doesn't calculate EVI if not present in the object.
}
\examples{
evi(dst_gpd(0, 1, 2))
evi(dst_gpd(0, 1, -1))
evi(dst_unif(0, 10))
evi(dst_norm(5, 5))
evi(stepdst(1:10))
}
