% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evi.R
\name{evi}
\alias{evi}
\title{Extreme Value Index}
\usage{
evi(x, ...)
}
\arguments{
\item{x}{Object to obtain EVI from.}

\item{...}{Other arguments to pass to specific methods.}
}
\value{
A single numeric.
}
\description{
Extreme Value Index
}
