% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn-generics.R, R/get_fn.dst.R
\name{get_quantile}
\alias{get_quantile}
\alias{get_quantile.dst}
\title{Get quantile function}
\usage{
get_quantile(object, ...)

\method{get_quantile}{dst}(object, tol = 1e-06, maxiter = 1000, ...)
}
\arguments{
\item{object}{Object of class "dst"}

\item{...}{Other arguments to pass to specific methods}

\item{tol}{tolerance}

\item{maxiter}{Maximum number of iterations}
}
\value{
a vectorized function of the quantile function.
}
\description{
Get quantile function
}
\seealso{
\code{\link{get_probfn}},
\code{\link{get_cdf}},
\code{\link{get_randfn}},
\code{\link{get_hazard}},
\code{\link{get_survival}}
}
