% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-graft.R
\name{graft_right}
\alias{graft_right}
\alias{graft_left}
\title{Graft Distribution}
\usage{
graft_right(distribution, graft, of, ..., include = FALSE)

graft_left(distribution, graft, of, ..., include = FALSE)
}
\arguments{
\item{distribution}{Base distribution}

\item{graft}{The distribution being grafted.}

\item{of}{Value on the real line where the graft is attached.}

\item{include}{Logical; include \code{of} in the base distribution?}
}
\value{
Graft distribution object, which is a special type of mixture
distribution.
}
\description{
Replace a distribution's tail.
\code{graft_left()} takes a base distribution and grafts a distribution
to the left of some value; \code{graft_right()} grafts a distribution
to the right. The distribution being grafted is trimmed at \code{of}
and fit to the base distribution also at that value.
}
\examples{
base <- distionary::dst_norm(0, 1)
q <- distionary::eval_quantile(base, at = 0.9)
right <- q + distionary::dst_gp(1, 0.3)
g <- graft_right(base, right, of = q)
# plot(g, "cdf", n = 1001)
# plot(base, "cdf", n = 1001, lty = 2, col = "green", add = TRUE)
}
