% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graft.R
\name{graft_right}
\alias{graft_right}
\alias{is_graft}
\alias{is.graft}
\title{Graft Distributions}
\usage{
graft_right(dst_left, dst_right, sep_y)

is_graft(object)

is.graft(object)
}
\arguments{
\item{dst_left, dst_right}{Distributions to connect}

\item{sep_y}{Value on the domain of the cdf to connect at.}

\item{object}{Object to be tested}
}
\value{
A grafted distribution object.
}
\description{
Replace the tail of a distribution.
\code{graft_right()} keeps the left cdf unchanged to the left of
sep_y, and makes a continuous connection with the right cdf
(rescaled as appropriate).
\code{graft_left()} keeps the right cdf unchanged to the right of
sep_y, and makes a continuous connection with the left cdf
(rescaled as appropriate),
}
