% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graft.R
\name{graft_right}
\alias{graft_right}
\alias{graft_left}
\alias{is_graft}
\title{Graft Distribution}
\usage{
graft_right(distribution, graft, breakpoint, include = FALSE)

graft_left(distribution, graft, breakpoint, include = FALSE)

is_graft(object)
}
\arguments{
\item{distribution}{Base distribution}

\item{graft}{The distribution being grafted.}

\item{breakpoint}{The location of the graft}

\item{include}{Logical; include the breakpoint in the distribution
being grafted?}

\item{object}{Object to be tested}
}
\value{
Graft distribution object, which is a special type of mixture
distribution.
}
\description{
Replace a distribution's tail.
\code{graft_left()} takes a base distribution and grafts a distribution
to the left of some breakpoint; \code{graft_right()} grafts a distribution
to the right. The distribution being grafted is sliced at a breakpoint
and fit to the base distribution also at that breakpoint.
}
\examples{
set.seed(1)
x <- stats::rcauchy(100)
base <- dst_empirical(x)
q <- eval_quantile(base, at = 0.9)
right <- dst_gpd(q, 5, 1)
g <- graft_right(base, right, breakpoint = q)
plot(g, "cdf", n = 1001, to = 34)
plot(base, "cdf", n = 1001, lty = 2, add = TRUE)
}
