% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graft.R
\name{graft_right}
\alias{graft_right}
\title{Replace the Tail of a Distribution}
\usage{
graft_right(left_dst, right_dst, sep_y)
}
\arguments{
\item{left_dst, right_dst}{Distributions to connect}

\item{sep_y}{Value on the domain of the cdf to connect at.}
}
\value{
Object of class "dst"
}
\description{
\code{graft_right()} keeps the left cdf unchanged to the left of
sep_y, and makes a continuous connection with the right cdf
(rescaled as appropriate).
\code{graft_left()} keeps the right cdf unchanged to the right of
sep_y, and makes a continuous connection with the left cdf
(rescaled as appropriate),
}
