% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframe.R, R/eval_hazard.R
\name{enframe_hazard}
\alias{enframe_hazard}
\alias{eval_hazard}
\alias{get_hazard}
\title{Hazard Function}
\usage{
enframe_hazard(..., at, arg_name = ".arg", fn_prefix = "hazard", sep = "_")

eval_hazard(object, at)

get_hazard(object)
}
\arguments{
\item{at}{Vector of values to evaluate the cdf at.}

\item{arg_name}{Name of the column containing the function arguments, in the
output data frame of \code{enframe_}.}

\item{object}{Distribution object.}
}
\value{
A vector of the evaluated hazard, in the case of \code{eval_}; a data
frame with both the argument and function evaluations, in the case of
\code{enframe_}; or a vectorized function representing the hazard, in the
case of \code{get_}.
}
\description{
Access a distribution's hazard function.
}
\examples{
d <- dst_unif(0, 4)
eval_hazard(d, at = 0:4)
enframe_hazard(d, at = 0:4)
hazard <- get_hazard(d)
hazard(0:4)
}
\seealso{
Other distributional representations: 
\code{\link{enframe_cdf}()},
\code{\link{enframe_chf}()},
\code{\link{enframe_density}()},
\code{\link{enframe_odds}()},
\code{\link{enframe_pmf}()},
\code{\link{enframe_quantile}()},
\code{\link{enframe_survival}()}
}
\concept{distributional representations}
