% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn-generics.R
\name{get_chf}
\alias{get_chf}
\alias{get_hazard}
\title{Hazard Function}
\usage{
get_chf(object)

get_hazard(object)
}
\arguments{
\item{object}{Distribution object}
}
\description{
Hazard and cumulative hazard functions.
}
