% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation-hazard.R
\name{eval_hazard}
\alias{eval_hazard}
\alias{get_hazard}
\alias{enframe_hazard}
\title{Hazard Function}
\usage{
eval_hazard(object, at)

get_hazard(object)

enframe_hazard(object, at, arg_name = ".arg", fn_name = ".hazard")
}
\arguments{
\item{object}{Object of class "dst"}

\item{at}{Vector of values to evaluate the hazard function at.}
}
\value{
Vector of the evaluated hazard function
}
\description{
Hazard Function
}
\seealso{
\code{\link{eval_quantile}},
\code{\link{eval_probfn}},
\code{\link{eval_cdf}},
\code{\link{eval_survival}}
}
