% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation-chf.R, R/representation-hazard.R
\name{get_chf}
\alias{get_chf}
\alias{eval_chf}
\alias{eval_hazard}
\alias{get_hazard}
\title{Hazard Function}
\usage{
get_chf(object)

eval_chf(object, at)

eval_hazard(object, at)

get_hazard(object)
}
\arguments{
\item{object}{Object of class "dst"}

\item{at}{Vector of values to evaluate the hazard function at.}
}
\value{
Vector of the evaluated hazard function
}
\description{
Hazard and cumulative hazard functions.

Evaluate hazard function or cumulative
hazard function (chf).
}
\seealso{
\code{\link{eval_quantile}},
\code{\link{eval_probfn}},
\code{\link{eval_cdf}},
\code{\link{realise}},
\code{\link{eval_survival}}
}
