% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_hazard.R
\name{eval_hazard}
\alias{eval_hazard}
\alias{enframe_hazard}
\title{Hazard Function}
\usage{
eval_hazard(distribution, at)

enframe_hazard(..., at, arg_name = ".arg", fn_prefix = "hazard", sep = "_")
}
\arguments{
\item{distribution}{A distribution, or possibly multiple
distributions in the case of \code{...}.}

\item{at}{Vector of values to evaluate the cdf at. Must be named when using
in \code{enframe_}.}

\item{...}{A distribution, or possibly multiple
distributions in the case of \code{...}.}

\item{arg_name}{For \code{enframe_}, name of the column containing
the function arguments.}

\item{fn_prefix}{For \code{enframe_}, name of the function to
appear in the column(s).}

\item{sep}{When \code{enframe}'ing more than one distribution, the
character that will be separating the \code{fn_name} and the distribution name.}
}
\value{
The evaluated hazard in vector form (for \code{eval_}) and data frame
or tibble form (for \code{enframe_}).
}
\description{
Access a distribution's hazard function.
}
\examples{
d <- dst_unif(0, 4)
eval_hazard(d, at = 0:4)
enframe_hazard(d, at = 0:4)
hazard <- get_hazard(d)
hazard(0:4)
}
\seealso{
Other distributional representations: 
\code{\link{eval_cdf}()},
\code{\link{eval_chf}()},
\code{\link{eval_density}()},
\code{\link{eval_odds}()},
\code{\link{eval_pmf}()},
\code{\link{eval_quantile}()},
\code{\link{eval_survival}()}
}
\concept{distributional representations}
