% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-invert.R
\name{invert}
\alias{invert}
\title{Inverse Distribution}
\usage{
invert(distribution)
}
\arguments{
\item{distribution}{A probability distribution of class \code{"dst"}.}
}
\value{
An inverse distribution.
}
\description{
Invert a distribution. Specifically, if \code{X} is a random
variable, \code{invert()} returns the distribution of \code{1 / X}.
The \code{/} operator can also be used to obtain the distribution.
}
\details{
Distributions must not have mass at 0, otherwise an error is returned
(i.e., \code{eval_pmf(distribution, at = 0)} is non-zero).

Simplifications are made in some situations, and will improve with future
versions of distplyr.
}
\examples{
invert(distionary::dst_norm(0, 1))
1 / (distionary::dst_pois(3.4) + 1)
}
\seealso{
\code{\link[=flip]{flip()}}
}
