% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_finite.R
\name{is_finite_dst}
\alias{is_finite_dst}
\alias{is.finite_dst}
\title{Is a distribution a finite distribution?}
\usage{
is_finite_dst(object)

is.finite_dst(object)
}
\arguments{
\item{object}{Object to check}
}
\description{
Is a distribution a finite distribution?
}
