% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_empirical.R, R/dst_finite.R
\name{is.empirical}
\alias{is.empirical}
\alias{is_empirical}
\alias{is_finite_dst}
\alias{is.finite_dst}
\title{Is a distribution a finite distribution?}
\usage{
is.empirical(object)

is_empirical(object)

is_finite_dst(object)

is.finite_dst(object)
}
\arguments{
\item{object}{Object to check}
}
\description{
There's no difference between checking whether a distribution
is finite or empirical -- the functions exist simply for
completeness.
}
