% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{make_dst_scale}
\alias{make_dst_scale}
\title{This helper function is not anticipating a negative constant.}
\usage{
make_dst_scale(e1)
}
\description{
This helper function is not anticipating a negative constant.
}
