% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-mix.R
\name{mix}
\alias{mix}
\alias{is_mix}
\alias{is.mix}
\title{Mixture Distributions}
\usage{
mix(..., weights = 1, na.rm = FALSE)

is_mix(object)

is.mix(object)
}
\arguments{
\item{...}{Distribution objects to mix, possibly also lists of distributions.}

\item{weights}{Vector of weights corresponding to the distributions;
or, single numeric for equal weights.}

\item{na.rm}{Remove distributions corresponding to \code{NA} weights?
Default is \code{FALSE}.}

\item{object}{Object to be tested}
}
\value{
A mixture distribution -- an empty distribution if any weights
are \code{NA} and \code{na.rm = FALSE}, the default.
}
\description{
Create a mixture distribution.
}
\examples{
a <- distionary::dst_norm(0, 1)
b <- distionary::dst_norm(5, 2)
m1 <- mix(a, b, weights = c(1, 4))
#plot(m1)
distionary::variable(m1)

c <- distionary::dst_empirical(0:6)
m2 <- mix(a, b, c, weights = c(0.2, 0.5, 0.3))
#plot(m2, n = 1001)
distionary::variable(m2)
}
