% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{mix}
\alias{mix}
\title{Mixture Distributions}
\usage{
mix(..., probs)
}
\arguments{
\item{...}{Distribution objects to mix.}

\item{probs}{Vector of probabilities corresponding to the distributions.}
}
\value{
A mixture distribution.
}
\description{
Create a mixture distribution.
}
