% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kurtosis.R, R/get_mean.R, R/get_skewness.R,
%   R/get_variance.R
\name{get_kurtosis_raw}
\alias{get_kurtosis_raw}
\alias{get_kurtosis_exc}
\alias{get_mean}
\alias{get_skewness}
\alias{get_variance}
\alias{get_sd}
\title{Moments of a Distribution}
\usage{
get_kurtosis_raw(object)

get_kurtosis_exc(object)

get_mean(object, ...)

get_skewness(object)

get_variance(object, ...)

get_sd(object, ...)
}
\arguments{
\item{object}{A distribution object.}

\item{...}{Arguments to pass to the \code{integrate()} function (if needed).}
}
\value{
A single numeric.
a <- dst_gpd(0, 1, 0.5)
b <- dst_unif(0, 1)
c <- dst_norm(3, 4)
get_mean(a)
get_variance(b)
get_kurtosis_raw(c)
get_kurtosis_exc(c)
get_mean(mix(a, b, probs = c(0.5, 0.5)))
}
\description{
Get common moment-related quantities of a
distribution: mean, variance, standard deviation (sd),
skewness, and kurtosis.
}
\details{
If there is no method associated with a subclass of
\code{object}, then moments are calculated using \code{stats::integrate()}
from the survival function.
}
