% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kurtosis.R, R/get_mean.R, R/get_skewness.R,
%   R/get_variance.R
\name{kurtosis_raw}
\alias{kurtosis_raw}
\alias{kurtosis_exc}
\alias{mean.dst}
\alias{skewness}
\alias{variance}
\alias{sd}
\title{Moments of a Distribution}
\usage{
kurtosis_raw(object)

kurtosis_exc(object)

\method{mean}{dst}(object, ...)

skewness(object)

variance(object, ...)

sd(object, ...)
}
\arguments{
\item{object}{A distribution object.}

\item{...}{Arguments to pass to the \code{integrate()} function (if needed).}
}
\value{
A single numeric.
}
\description{
Get common moment-related quantities of a
distribution: mean, variance, standard deviation (sd),
skewness, and kurtosis.
}
\details{
If there is no method associated with a subclass of
\code{object}, then moments are calculated using \code{stats::integrate()}
from the survival function.
}
\examples{
a <- dst_gpd(0, 1, 0.5)
b <- dst_unif(0, 1)
c <- dst_norm(3, 4)
mean(a)
variance(b)
kurtosis_raw(c)
kurtosis_exc(c)
mean(mix(a, b, probs = c(0.5, 0.5)))
}
