% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_discrete.R
\name{next_discrete}
\alias{next_discrete}
\title{Return neighbouring discrete values from a reference}
\usage{
next_discrete(object, from, n = 1L, include_from = FALSE, ...)
}
\arguments{
\item{object}{Distribution}

\item{from}{Reference value.}

\item{n}{Number of discrete values to find.}

\item{include_from}{Logical; should the \code{from} value be included
in the query?}
}
\value{
Vector of all available discrete points satisfying the query.
If infinite values satify the query, an error is thrown; if less
values are available than asked via \code{n}, only those values are returned.
}
\description{
Finds the next \code{n} largest/smallest discrete values in a distribution
from some reference point.
}
\examples{
next_discrete(dst_pois(1), from = 1.3)
prev_discrete(dst_pois(1), from = 3, n = 10)
next_discrete(dst_norm(0, 1), from = 1.3, n = 4)
}
