% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-quantile.R
\name{get_higher}
\alias{get_higher}
\alias{get_lower}
\title{Find an Encapsulating Range of Quantiles}
\usage{
get_higher(cdf, level)

get_lower(cdf, level)
}
\arguments{
\item{cdf}{Function representing the cdf}

\item{level}{Quantile level defining the}
}
\description{
Given a quantile level, these functions
will find a quantile that evaluates to
be higher/lower than the provided quantile level.
Intended for internal use only.
}
