% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation-probfn.R
\name{eval_probfn}
\alias{eval_probfn}
\alias{get_probfn}
\title{Probability density/mass function}
\usage{
eval_probfn(object, at)

get_probfn(object)
}
\arguments{
\item{object}{Object of class "dst"}

\item{at}{Vector of values to evaluate the density or mass function at}
}
\value{
Vector of density/mass values
}
\description{
Probability density/mass function
}
\seealso{
\code{\link{eval_quantile}},
\code{\link{eval_cdf}},
\code{\link{realise}},
\code{\link{eval_hazard}},
\code{\link{eval_survival}}
}
