% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation-quantile.R
\name{eval_quantile}
\alias{eval_quantile}
\alias{get_quantile}
\alias{get_quantile.dst}
\alias{eval_quantile.dst}
\alias{enframe_quantile}
\title{Distribution Quantiles}
\usage{
eval_quantile(object, at, ...)

get_quantile(object, ...)

\method{get_quantile}{dst}(object, tol = 1e-06, maxiter = 1000, ...)

\method{eval_quantile}{dst}(object, at, tol = 1e-06, maxiter = 1000, ...)

enframe_quantile(object, at, arg_name = ".arg", fn_name = ".quantile")
}
\arguments{
\item{object}{Object of class "dst"}

\item{at}{Vector of quantile levels (probabilities).}

\item{...}{Other arguments to pass to specific methods}

\item{tol}{tolerance}

\item{maxiter}{Maximum number of iterations}
}
\value{
Vector of quantiles
}
\description{
Distribution Quantiles
}
\seealso{
\code{\link{eval_probfn}},
\code{\link{eval_cdf}},
\code{\link{eval_hazard}},
\code{\link{eval_survival}}
}
