% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_degen.R, R/dst_finite.R, R/dst_gpd.R,
%   R/dst_norm.R, R/dst_unif.R, R/range.r
\name{range.degenerate}
\alias{range.degenerate}
\alias{range.finite}
\alias{range.gpd}
\alias{range.norm}
\alias{range.unif}
\alias{range}
\title{Range of Distribution}
\usage{
\method{range}{degenerate}(x, ...)

\method{range}{finite}(x, ...)

\method{range}{gpd}(x, ...)

\method{range}{norm}(x, ...)

\method{range}{unif}(x, ...)

range(x, ...)
}
\description{
Generic function to find range of distributions
}
