% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities-dst_gpd.R, R/quantities-dst_lnorm.R,
%   R/quantities-dst_norm.R, R/quantities-dst_pois.R, R/quantities-dst_unif.R,
%   R/quantities-finite.R, R/range.r
\name{range.gpd}
\alias{range.gpd}
\alias{range.lnorm}
\alias{range.norm}
\alias{range.pois}
\alias{range.unif}
\alias{range.finite}
\alias{range.dst}
\title{Range of Distribution}
\usage{
\method{range}{gpd}(x, ...)

\method{range}{lnorm}(x, ...)

\method{range}{norm}(x, ...)

\method{range}{pois}(x, ...)

\method{range}{unif}(x, ...)

\method{range}{finite}(x, ...)

\method{range}{dst}(x, ...)
}
\arguments{
\item{x}{Object to compute range}

\item{...}{Other arguments to pass to specific methods. Currently not used.}
}
\description{
Range returns a 2 index vector with the 0th index
containing the minimum value, and the 1st index containing the maximum value
for a given distribution; currently, the ellipsis argument has no
functionality.
}
\details{
If there are no methods for the x's class, the range is calculated
using eval_quantile at 0 and at 1
}
\examples{
a <- dst_gpd(0, 1, 0.5)
b <- dst_unif(0, 1)
c <- dst_norm(3, 4)
range(a)
range(b)
range(c)
}
