% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realise.R
\name{realise}
\alias{realise}
\alias{realize}
\title{Generate a Sample from a Distribution}
\usage{
realise(object, n = 1, ...)

realize(object, n = 1, ...)
}
\arguments{
\item{object}{Object of class "dst"}

\item{n}{Number of observations to generate}

\item{...}{Other arguments to pass to \code{\link{eval_quantile}}}
}
\value{
Vector of independent values drawn from the inputted distribution
}
\description{
Generate a Sample from a Distribution
}
