% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variance.R
\name{sd.default}
\alias{sd.default}
\title{Access to the original sd function}
\usage{
\method{sd}{default}(object, ...)
}
\arguments{
\item{object}{Object for which to compute the standard deviation}

\item{...}{Other arguments to pass to \code{stats::sd()}.}
}
\description{
Since distplyr makes \code{sd()} a generic function, this
default method dispatches the \code{stats::sd()} function
whenever \code{sd()} is called on a non-distribution object.
}
\seealso{
\code{stats::sd}
}
