% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-trim.R
\name{trim_left}
\alias{trim_left}
\alias{trim_right}
\title{Trim (condition) a distribution}
\usage{
trim_left(distribution, of, ..., include = TRUE)

trim_right(distribution, of, ..., include = TRUE)
}
\arguments{
\item{distribution}{Distribution to trim.}

\item{of}{Value on the real line defining where to trim (single numeric).}

\item{...}{Currently unused.}

\item{include}{Logical; should \code{of} be removed from the support as well?
This is only realistically relevant if \code{of} has a non-zero probability
of occurrence.}
}
\value{
A conditional distribution.
}
\description{
\code{trim_left()} removes probability to the left of some value,
conditioning the random variable to be bigger than that value.
\code{trim_right()} does the opposite: removes probability to the right,
conditioning to be smaller than that value.
}
\examples{
d <- distionary::dst_norm(0, 1)
d <- trim_left(d, -2)
d <- trim_right(d, 2)
distionary::enframe_cdf(d, at = -3:3)

d <- distionary::dst_pois(3)
distionary::eval_pmf(trim_left(d, 5), at = 5)
distionary::eval_pmf(trim_left(d, 5, include = FALSE), at = 5)
}
