% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-trim.R
\name{trim_left}
\alias{trim_left}
\alias{trim_right}
\title{Trim (condition) a distribution}
\usage{
trim_left(distribution, of, ..., include = TRUE)

trim_right(distribution, of, ..., include = TRUE)
}
\arguments{
\item{distribution}{Distribution to trim.}

\item{of}{Value on the real line defining where to trim (single numeric).}

\item{...}{Currently unused.}

\item{include}{Logical; should \code{of} be removed from the support as well?
This is only realistically relevant if \code{of} has a non-zero probability
of occurrence.}
}
\value{
A conditional distribution.
}
\description{
\code{trim_left()} removes probability to the left of some value,
conditioning the random variable to be bigger than that value.
\code{trim_right()} does the opposite: removes probability to the right,
conditioning to be smaller than that value.
}
\examples{
distionary::dst_norm(0, 1) |>
  trim_left(-2) |>
  trim_right(2) |>
  distionary::enframe_cdf(at = -3:3)

d <- distionary::dst_pois(3)
d |>
  trim_left(5) |>
  distionary::eval_pmf(at = 5)
d |>
  trim_left(5, include = FALSE) |>
  distionary::eval_pmf(at = 5)
}
