% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{variable}
\alias{variable}
\title{Random Variable of a Distribution}
\usage{
variable(distribution)
}
\arguments{
\item{distribution}{Distribution.}
}
\value{
As of now, the type of variable that it is -- one
of "continuous", "discrete", or "mixed".
}
\description{
Information about a random variable of a distribution.
}
